Split InterceptorCodeGenerator into partial class files by domain concern
Closes #2
