+++ src/Quarry.Generator/Generation/InterceptorCodeGenerator.cs
@@ -18,7 +18,7 @@ namespace Quarry.Generators.Generation;
 /// to generated code that provides compile-time SQL generation and typed reader delegates.
 /// All interceptors use the 'file' scope modifier to prevent visibility to end users.
 /// </remarks>
-internal static class InterceptorCodeGenerator
+internal static partial class InterceptorCodeGenerator
 {
     /// <summary>
     /// Generates the interceptors file for a context.
@@ -373,3044 +373,5 @@ internal static class InterceptorCodeGenerator
         sb.AppendLine("        public InterceptsLocationAttribute(int version, string data) { }");
         sb.AppendLine("    }");
         sb.AppendLine("}");
-    }
-
-    /// <summary>
-    /// Extracts the namespace from a syntax node's containing file.
-    /// </summary>
-    private static string? GetNamespaceFromFilePath(Microsoft.CodeAnalysis.SyntaxNode? syntaxNode)
-    {
-        if (syntaxNode?.SyntaxTree == null)
-            return null;
-
-        var root = syntaxNode.SyntaxTree.GetRoot();
-
-        // Look for file-scoped namespace
-        var fileScopedNamespace = root.DescendantNodes()
-            .OfType<Microsoft.CodeAnalysis.CSharp.Syntax.FileScopedNamespaceDeclarationSyntax>()
-            .FirstOrDefault();
-        if (fileScopedNamespace != null)
-        {
-            return fileScopedNamespace.Name.ToString();
-        }
-
-        // Look for block-scoped namespace containing the invocation
-        var namespaceDecl = syntaxNode.Ancestors()
-            .OfType<Microsoft.CodeAnalysis.CSharp.Syntax.NamespaceDeclarationSyntax>()
-            .FirstOrDefault();
-        if (namespaceDecl != null)
-        {
-            return namespaceDecl.Name.ToString();
-        }
-
-        // Fallback: look for any namespace in the file
-        var anyNamespace = root.DescendantNodes()
-            .OfType<Microsoft.CodeAnalysis.CSharp.Syntax.NamespaceDeclarationSyntax>()
-            .FirstOrDefault();
-        if (anyNamespace != null)
-        {
-            return anyNamespace.Name.ToString();
-        }
-
-        return null;
-    }
-
-    /// <summary>
-    /// Converts an absolute file path to a shorter relative-style path for display in comments.
-    /// Returns a path starting from a recognized project marker, or the last few path segments.
-    /// </summary>
-    private static string GetRelativePath(string filePath)
-    {
-        if (string.IsNullOrEmpty(filePath))
-            return filePath;
-
-        // Common project root markers to look for
-        string[] markers = ["src", "test", "tests", "lib", "samples", "examples", "benchmark", "benchmarks"];
-
-        foreach (var marker in markers)
-        {
-            // Try with backslash (Windows)
-            var index = filePath.IndexOf(marker + "\\", StringComparison.OrdinalIgnoreCase);
-            if (index >= 0)
-                return filePath.Substring(index);
-
-            // Try with forward slash (Unix/normalized)
-            index = filePath.IndexOf(marker + "/", StringComparison.OrdinalIgnoreCase);
-            if (index >= 0)
-                return filePath.Substring(index);
-        }
-
-        // No marker found - return last 3 path segments for context
-        var segments = filePath.Split(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
-        if (segments.Length <= 3)
-            return filePath;
-
-        var lastSegments = new string[3];
-        Array.Copy(segments, segments.Length - 3, lastSegments, 0, 3);
-        return string.Join(Path.DirectorySeparatorChar.ToString(), lastSegments);
-    }
-
-    /// <summary>
-    /// Extracts the namespace from a fully qualified type name.
-    /// </summary>
-    private static string? GetNamespaceFromTypeName(string typeName)
-    {
-        // Handle global:: prefix
-        if (typeName.StartsWith("global::"))
-        {
-            typeName = typeName.Substring(8);
-        }
-
-        // Find the last dot that's not within angle brackets (generic types)
-        int depth = 0;
-        int lastDotIndex = -1;
-
-        for (int i = typeName.Length - 1; i >= 0; i--)
-        {
-            char c = typeName[i];
-            if (c == '>')
-                depth++;
-            else if (c == '<')
-                depth--;
-            else if (c == '.' && depth == 0)
-            {
-                lastDotIndex = i;
-                break;
-            }
-        }
-
-        if (lastDotIndex <= 0)
-            return null;
-
-        return typeName.Substring(0, lastDotIndex);
-    }
-
-    /// <summary>
-    /// Checks if a clause interceptor should be skipped because the clause could not be translated.
-    /// When skipped, the original runtime method runs instead of a silent no-op fallback.
-    /// </summary>
-    private static bool ShouldSkipNonTranslatableClause(UsageSiteInfo site)
-    {
-        switch (site.Kind)
-        {
-            case InterceptorKind.Where:
-            case InterceptorKind.DeleteWhere:
-            case InterceptorKind.UpdateWhere:
-                return site.ClauseInfo == null || !site.ClauseInfo.IsSuccess;
-
-            case InterceptorKind.OrderBy:
-            case InterceptorKind.ThenBy:
-                if (site.ClauseInfo is OrderByClauseInfo orderByInfo && orderByInfo.IsSuccess)
-                    return false;
-                return site.ClauseInfo == null || !site.ClauseInfo.IsSuccess;
-
-            case InterceptorKind.GroupBy:
-            case InterceptorKind.Having:
-                return site.ClauseInfo == null || !site.ClauseInfo.IsSuccess;
-
-            case InterceptorKind.Set:
-            case InterceptorKind.UpdateSet:
-                if (site.ClauseInfo is SetClauseInfo setInfo && setInfo.IsSuccess)
-                    return false;
-                return site.ClauseInfo == null || !site.ClauseInfo.IsSuccess;
-
-            case InterceptorKind.UpdateSetPoco:
-                return site.UpdateInfo == null || site.UpdateInfo.Columns.Count == 0;
-
-            case InterceptorKind.InsertExecuteNonQuery:
-            case InterceptorKind.InsertExecuteScalar:
-                return site.InsertInfo == null || site.InsertInfo.Columns.Count == 0;
-
-            default:
-                return false;
-        }
-    }
-
-    /// <summary>
-    /// Checks if a Select interceptor should be skipped entirely (not emitted).
-    /// </summary>
-    private static bool ShouldSkipSelectInterceptor(UsageSiteInfo site)
-    {
-        var projection = site.ProjectionInfo;
-        if (projection == null)
-            return false; // Let the fallback path handle it
-
-        // Skip anonymous types - they should have been rejected by QRY014
-        if (projection.Kind == ProjectionKind.Anonymous)
-            return true;
-
-        // Skip unknown or failed projections
-        if (projection.Kind == ProjectionKind.Unknown)
-            return true;
-
-        // Skip non-optimal paths
-        if (!projection.IsOptimalPath)
-            return true;
-
-        // Skip tuple projections with invalid result types
-        // This can happen when the entity types are generated by the same source generator
-        // and the semantic model can't fully resolve the tuple element types
-        if (projection.Kind == ProjectionKind.Tuple)
-        {
-            var resultType = projection.ResultTypeName;
-            if (string.IsNullOrWhiteSpace(resultType) ||
-                !resultType.StartsWith("(") ||
-                resultType.Contains("( ") ||
-                resultType.Contains(", )") ||
-                ContainsUnresolvedTupleElement(resultType))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Checks if a tuple type string contains unresolved elements (missing types).
-    /// </summary>
-    private static bool ContainsUnresolvedTupleElement(string tupleType)
-    {
-        // Remove parentheses
-        var inner = tupleType.Substring(1, tupleType.Length - 2);
-        var parts = inner.Split(',');
-
-        foreach (var part in parts)
-        {
-            var trimmed = part.Trim();
-            if (string.IsNullOrEmpty(trimmed))
-                return true;
-
-            // Check if this looks like just a name without a type
-            // Valid: "int UserId", "string? Name"
-            // Invalid: "UserId", "Name"
-            var spaceIdx = trimmed.IndexOf(' ');
-            if (spaceIdx < 0)
-            {
-                // Single word - must be a type like "int" or "DateTime"
-                // If it starts with uppercase and isn't a known type, it's likely just a name
-                if (trimmed.Length > 0 && char.IsUpper(trimmed[0]) && !IsKnownUppercaseTypeName(trimmed))
-                {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Checks if a name is a known .NET type that starts with uppercase.
-    /// </summary>
-    private static bool IsKnownUppercaseTypeName(string name)
-    {
-        var baseName = name.TrimEnd('?');
-        return baseName switch
-        {
-            "Boolean" or "Byte" or "SByte" or "Int16" or "UInt16" or
-            "Int32" or "UInt32" or "Int64" or "UInt64" or
-            "Single" or "Double" or "Decimal" or "String" or "Char" or
-            "DateTime" or "DateTimeOffset" or "TimeSpan" or "DateOnly" or "TimeOnly" or
-            "Guid" or "Object" => true,
-            _ => false
-        };
-    }
-
-    /// <summary>
-    /// Sanitizes a tuple type name by stripping element names and keeping only types.
-    /// For example, "(string Status, int)" becomes "(string, int)".
-    /// This prevents element names from being emitted as types in interceptor signatures
-    /// when the semantic model produces unresolved element types.
-    /// </summary>
-    private static string SanitizeTupleResultType(string tupleType)
-    {
-        // Delegate to the full implementation
-        return SanitizeTupleResultTypeCore(tupleType);
-    }
-
-    /// <summary>
-    /// Resolves the best available result type for execution interceptors.
-    /// Prefers the semantic model's type (most qualified) but falls back to
-    /// enriched ProjectionInfo type for tuples, single-column, and empty results.
-    /// </summary>
-    private static string? ResolveExecutionResultType(
-        string? siteResultType,
-        string? chainResultType,
-        ProjectionInfo? projectionInfo)
-    {
-        // For non-tuple, non-empty results, prefer the semantic model's type (most qualified)
-        if (!string.IsNullOrEmpty(siteResultType)
-            && !siteResultType!.Contains("(") && !siteResultType.Contains("ValueTuple"))
-            return siteResultType;
-        if (!string.IsNullOrEmpty(chainResultType)
-            && !chainResultType!.Contains("(") && !chainResultType.Contains("ValueTuple"))
-            return chainResultType;
-
-        // For tuples, empty results, or broken types: use enriched projection type
-        var projResult = projectionInfo?.ResultTypeName;
-        if (!string.IsNullOrEmpty(projResult))
-            return projResult;
-
-        // Last resort: return whatever we have (may be empty/broken)
-        return string.IsNullOrEmpty(siteResultType) ? chainResultType : siteResultType;
-    }
-
-    /// <summary>
-    /// Strips tuple element names from a C# tuple type, keeping only the type parts.
-    /// For example, "(string Status, int)" becomes "(string, int)".
-    /// </summary>
-    private static string SanitizeTupleResultTypeCore(string tupleType)
-    {
-        if (!tupleType.StartsWith("(") || !tupleType.EndsWith(")"))
-            return tupleType;
-
-        var inner = tupleType.Substring(1, tupleType.Length - 2);
-        var parts = inner.Split(',');
-        var sanitized = new string[parts.Length];
-
-        for (int i = 0; i < parts.Length; i++)
-        {
-            var trimmed = parts[i].Trim();
-            var spaceIdx = trimmed.IndexOf(' ');
-            if (spaceIdx > 0)
-            {
-                var typePart = trimmed.Substring(0, spaceIdx);
-                var namePart = trimmed.Substring(spaceIdx + 1);
-
-                // If the "type" part is empty/whitespace or looks like a name without a type,
-                // then the whole token is just a name — use "object" as fallback
-                if (string.IsNullOrWhiteSpace(typePart))
-                {
-                    sanitized[i] = "object";
-                }
-                else
-                {
-                    // Strip the element name, keep only the type
-                    sanitized[i] = typePart;
-                }
-            }
-            else
-            {
-                // Single word — could be a type ("int") or a misplaced name ("Status")
-                if (trimmed.Length > 0 && char.IsUpper(trimmed[0]) &&
-                    !trimmed.Contains('.') && !IsKnownUppercaseTypeName(trimmed))
-                {
-                    // Likely a name, not a type — use "object"
-                    sanitized[i] = "object";
-                }
-                else
-                {
-                    sanitized[i] = trimmed;
-                }
-            }
-        }
-
-        return $"({string.Join(", ", sanitized)})";
-    }
-
-    /// <summary>
-    /// Generates an interceptor method for a usage site.
-    /// </summary>
-    private static void GenerateInterceptorMethod(
-        StringBuilder sb,
-        UsageSiteInfo site,
-        List<CachedExtractorField> staticFields,
-        Dictionary<string, PrebuiltChainInfo> chainLookup,
-        Dictionary<string, int> clauseBitMap,
-        Dictionary<string, PrebuiltChainInfo> chainClauseLookup,
-        HashSet<string> firstClauseIds)
-    {
-        // Check for skippable Select sites BEFORE emitting the attribute
-        if (site.Kind == InterceptorKind.Select && ShouldSkipSelectInterceptor(site))
-        {
-            return; // Don't emit anything for this site
-        }
-
-        // Limit/Offset/Distinct are tracked for chain analysis only — they don't need
-        // their own interceptor methods. The builder methods are simple property setters.
-        if (site.Kind is InterceptorKind.Limit or InterceptorKind.Offset or InterceptorKind.Distinct)
-        {
-            return;
-        }
-
-        // For execution interceptors: only emit if we have a pre-built chain for this site
-        // AND we can resolve the result type. Otherwise skip — the built-in execution methods
-        // work correctly via the Select interceptor.
-        if (site.Kind is InterceptorKind.ExecuteFetchAll or InterceptorKind.ExecuteFetchFirst
-            or InterceptorKind.ExecuteFetchFirstOrDefault or InterceptorKind.ExecuteFetchSingle
-            or InterceptorKind.ToAsyncEnumerable)
-        {
-            if (!chainLookup.TryGetValue(site.UniqueId, out var chain))
-                return;
-            // Skip chains with unmatched fluent methods (Limit, Offset, AddWhereClause, etc.)
-            // whose effects are invisible to the pre-built SQL.
-            if (chain.Analysis.UnmatchedMethodNames != null)
-                return;
-            // Resolve the best available result type. The site/chain result types come from the
-            // semantic model and may be empty or broken for source-generated entity types.
-            // The enriched ProjectionInfo.ResultTypeName is rebuilt from entity metadata and is
-            // authoritative for tuples and single-column projections.
-            var rawResult = ResolveExecutionResultType(site.ResultTypeName, chain.ResultTypeName, chain.ProjectionInfo);
-            if (string.IsNullOrEmpty(rawResult))
-                return;
-            // Must have a reader delegate
-            if (chain.ReaderDelegateCode == null)
-                return;
-            // Skip chains with projections that have SqlExpression set alongside ColumnName.
-            // CompileTimeSqlBuilder uses SqlExpression (e.g., "LOWER(col)") but the runtime path
-            // uses ColumnName and quotes it. These produce different SQL.
-            // Only allow execution interceptors when the SQL will be byte-identical.
-            if (chain.ProjectionInfo != null)
-            {
-                var hasAmbiguousColumns = chain.ProjectionInfo.Columns.Any(c =>
-                    c.SqlExpression != null && !string.IsNullOrEmpty(c.ColumnName));
-                if (hasAmbiguousColumns)
-                    return;
-            }
-        }
-        else if (site.Kind is InterceptorKind.ExecuteScalar)
-        {
-            if (!chainLookup.TryGetValue(site.UniqueId, out var scalarChain))
-                return;
-            if (scalarChain.Analysis.UnmatchedMethodNames != null)
-                return;
-            // ExecuteScalar doesn't need a reader delegate — just pre-built SQL
-            var rawScalarResult = ResolveExecutionResultType(site.ResultTypeName, scalarChain.ResultTypeName, scalarChain.ProjectionInfo);
-            if (string.IsNullOrEmpty(rawScalarResult))
-                return;
-        }
-        else if (site.Kind is InterceptorKind.ExecuteNonQuery)
-        {
-            if (!chainLookup.TryGetValue(site.UniqueId, out var nqChain))
-                return;
-            // Validate that all SQL variants are non-empty and well-formed
-            if (nqChain.SqlMap.Values.Any(v => string.IsNullOrWhiteSpace(v.Sql)
-                || (nqChain.QueryKind == QueryKind.Update && v.Sql.Contains("SET  "))))
-                return;
-        }
-        else if (site.Kind is InterceptorKind.ToSql)
-        {
-            if (!chainLookup.TryGetValue(site.UniqueId, out var toSqlChain))
-                return;
-            if (toSqlChain.Analysis.UnmatchedMethodNames != null)
-                return;
-        }
-
-        // Skip clause interceptors where the clause could not be translated to SQL.
-        // The original runtime method will run instead. A QRY019 diagnostic is reported separately.
-        if (ShouldSkipNonTranslatableClause(site))
-        {
-            return;
-        }
-
-        var methodName = $"{site.MethodName}_{site.UniqueId}";
-
-        // Determine chain analysis status for remarks
-        PrebuiltChainInfo? chainForRemarks = null;
-        chainLookup.TryGetValue(site.UniqueId, out chainForRemarks);
-        if (chainForRemarks == null)
-            chainClauseLookup.TryGetValue(site.UniqueId, out chainForRemarks);
-
-        // XML documentation
-        sb.AppendLine($"    /// <summary>");
-        sb.AppendLine($"    /// Intercepts {site.MethodName}() call at {GetRelativePath(site.FilePath)}:{site.Line}:{site.Column}");
-        sb.AppendLine($"    /// </summary>");
-        if (chainForRemarks != null)
-            sb.AppendLine($"    /// <remarks>Chain: Fully Analyzed ({chainForRemarks.Analysis.Tier})</remarks>");
-        else
-            sb.AppendLine($"    /// <remarks>Chain: Not Analyzed (standalone interceptor)</remarks>");
-
-        // InterceptsLocation attribute - use new format with version and data if available
-        if (!string.IsNullOrEmpty(site.InterceptableLocationData))
-        {
-            sb.AppendLine($"    [InterceptsLocation({site.InterceptableLocationVersion}, \"{site.InterceptableLocationData}\")]");
-        }
-        else
-        {
-            // Fallback: skip this site if we don't have valid location data
-            sb.AppendLine($"    // WARNING: Could not generate InterceptsLocation for this call site");
-            return;
-        }
-
-        // Check if this is on a joined builder type
-        var isJoinedBuilder = site.JoinedEntityTypeNames != null && site.JoinedEntityTypeNames.Count >= 2;
-
-        // Look up clause bitmask bit index for conditional clause interceptors
-        clauseBitMap.TryGetValue(site.UniqueId, out var rawBit);
-        int? clauseBit = clauseBitMap.ContainsKey(site.UniqueId) ? rawBit : null;
-
-        // Resolve prebuilt chain membership for clause sites
-        chainClauseLookup.TryGetValue(site.UniqueId, out var prebuiltClauseChain);
-        var isFirstClauseInChain = prebuiltClauseChain != null && firstClauseIds.Contains(site.UniqueId);
-
-        // Generate method based on kind
-        switch (site.Kind)
-        {
-            case InterceptorKind.Select:
-                if (isJoinedBuilder)
-                    GenerateJoinedSelectInterceptor(sb, site, methodName, prebuiltClauseChain, isFirstClauseInChain);
-                else
-                    GenerateSelectInterceptor(sb, site, methodName, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.Where:
-                if (isJoinedBuilder)
-                    GenerateJoinedWhereInterceptor(sb, site, methodName, staticFields, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                else
-                    GenerateWhereInterceptor(sb, site, methodName, staticFields, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.OrderBy:
-            case InterceptorKind.ThenBy:
-                if (isJoinedBuilder)
-                    GenerateJoinedOrderByInterceptor(sb, site, methodName, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                else
-                    GenerateOrderByInterceptor(sb, site, methodName, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.GroupBy:
-                GenerateGroupByInterceptor(sb, site, methodName, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.Having:
-                GenerateHavingInterceptor(sb, site, methodName, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.Set:
-                GenerateSetInterceptor(sb, site, methodName, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.Join:
-            case InterceptorKind.LeftJoin:
-            case InterceptorKind.RightJoin:
-                GenerateJoinInterceptor(sb, site, methodName);
-                break;
-
-            case InterceptorKind.ExecuteFetchAll:
-            case InterceptorKind.ExecuteFetchFirst:
-            case InterceptorKind.ExecuteFetchFirstOrDefault:
-            case InterceptorKind.ExecuteFetchSingle:
-            case InterceptorKind.ExecuteScalar:
-            case InterceptorKind.ToAsyncEnumerable:
-                if (chainLookup.TryGetValue(site.UniqueId, out var selectChain))
-                {
-                    if (selectChain.IsJoinChain)
-                        GeneratePrebuiltJoinExecutionInterceptor(sb, site, methodName, selectChain);
-                    else
-                        GeneratePrebuiltSelectExecutionInterceptor(sb, site, methodName, selectChain);
-                }
-                break;
-
-            case InterceptorKind.ExecuteNonQuery:
-                if (chainLookup.TryGetValue(site.UniqueId, out var nonQueryChain))
-                    GeneratePrebuiltNonQueryExecutionInterceptor(sb, site, methodName, nonQueryChain);
-                break;
-
-            case InterceptorKind.ToSql:
-                if (chainLookup.TryGetValue(site.UniqueId, out var toSqlChain))
-                    GeneratePrebuiltToSqlInterceptor(sb, site, methodName, toSqlChain);
-                break;
-
-            case InterceptorKind.DeleteWhere:
-                GenerateDeleteWhereInterceptor(sb, site, methodName, staticFields, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.UpdateSet:
-                GenerateUpdateSetInterceptor(sb, site, methodName, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.UpdateSetPoco:
-                GenerateUpdateSetPocoInterceptor(sb, site, methodName);
-                break;
-
-            case InterceptorKind.UpdateWhere:
-                GenerateUpdateWhereInterceptor(sb, site, methodName, staticFields, clauseBit, prebuiltClauseChain, isFirstClauseInChain);
-                break;
-
-            case InterceptorKind.InsertExecuteNonQuery:
-                GenerateInsertExecuteNonQueryInterceptor(sb, site, methodName);
-                break;
-
-            case InterceptorKind.InsertExecuteScalar:
-                GenerateInsertExecuteScalarInterceptor(sb, site, methodName);
-                break;
-
-            case InterceptorKind.InsertToSql:
-                GenerateInsertToSqlInterceptor(sb, site, methodName);
-                break;
-
-            case InterceptorKind.RawSqlAsync:
-                GenerateRawSqlAsyncInterceptor(sb, site, methodName);
-                break;
-
-            case InterceptorKind.RawSqlScalarAsync:
-                GenerateRawSqlScalarAsyncInterceptor(sb, site, methodName);
-                break;
-
-            default:
-                GeneratePlaceholderInterceptor(sb, site, methodName);
-                break;
-        }
-
-        sb.AppendLine();
-    }
-
-    /// <summary>
-    /// Generates a Select() interceptor with column list and typed reader delegate.
-    /// </summary>
-    private static void GenerateSelectInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var projection = site.ProjectionInfo;
-
-        // Simplified prebuilt chain path: AsProjected instead of SelectWithReader
-        if (prebuiltChain != null && projection != null && projection.IsOptimalPath && projection.Columns.Count > 0)
-        {
-            var resultType = GetShortTypeName(projection.ResultTypeName);
-            var thisType = site.BuilderTypeName;
-            var concreteType = ToConcreteTypeName(thisType);
-            sb.AppendLine($"    public static {thisType}<{entityType}, {resultType}> {methodName}(");
-            sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-            sb.AppendLine($"        Func<{entityType}, {resultType}> _)");
-            sb.AppendLine($"    {{");
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        __b.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            sb.AppendLine($"        return __b.AsProjected<{resultType}>();");
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        // If we have projection info and it's optimal, generate the full interceptor
-        if (projection != null &&
-            projection.IsOptimalPath &&
-            projection.Columns.Count > 0)
-        {
-            GenerateOptimalSelectInterceptor(sb, site, methodName, entityType, projection);
-        }
-        else
-        {
-            // Fallback - delegate to the original method
-            GenerateFallbackSelectInterceptor(sb, site, methodName, entityType);
-        }
-    }
-
-    /// <summary>
-    /// Generates an optimal Select() interceptor with compile-time column binding.
-    /// </summary>
-    private static void GenerateOptimalSelectInterceptor(
-        StringBuilder sb,
-        UsageSiteInfo site,
-        string methodName,
-        string entityType,
-        ProjectionInfo projection)
-    {
-        // Generate column list SQL using the dialect resolved from the owning context
-        var columnList = ReaderCodeGenerator.GenerateColumnList(projection, site.Dialect ?? SqlDialect.PostgreSQL);
-
-        // Generate column names array
-        var columnNames = ReaderCodeGenerator.GenerateColumnNamesArray(projection);
-
-        // Generate reader delegate
-        var readerDelegate = ReaderCodeGenerator.GenerateReaderDelegate(projection, entityType);
-
-        // At this point, skippable cases (anonymous, tuple, unknown) have already been
-        // filtered out by ShouldSkipSelectInterceptor, so we can generate the interceptor
-        {
-            var resultType = GetShortTypeName(projection.ResultTypeName);
-            var thisType = site.BuilderTypeName;
-            var concreteType = ToConcreteTypeName(thisType);
-            sb.AppendLine($"    public static {thisType}<{entityType}, {resultType}> {methodName}(");
-            sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-            sb.AppendLine($"        Func<{entityType}, {resultType}> _)");
-            sb.AppendLine($"    {{");
-            sb.AppendLine($"        // Generated column list: {columnList}");
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-            sb.AppendLine($"        return __b.SelectWithReader(");
-            sb.AppendLine($"            {columnNames},");
-            sb.AppendLine($"            {readerDelegate});");
-            sb.AppendLine($"    }}");
-        }
-    }
-
-    /// <summary>
-    /// Generates a fallback Select() interceptor that delegates to the original method.
-    /// </summary>
-    private static void GenerateFallbackSelectInterceptor(
-        StringBuilder sb,
-        UsageSiteInfo site,
-        string methodName,
-        string entityType)
-    {
-        // Interceptor arity must match the combined class + method type parameters.
-        // QueryBuilder<T>.Select<TResult> has total arity 2 (T + TResult).
-        // Use arity 2 so the interceptor can match any Select<TResult> call.
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-        sb.AppendLine($"    public static {thisType}<T, TResult> {methodName}<T, TResult>(");
-        sb.AppendLine($"        this {thisType}<T> builder,");
-        sb.AppendLine($"        Func<T, TResult> selector) where T : class");
-        sb.AppendLine($"    {{");
-        sb.AppendLine($"        // Fallback path - projection not fully analyzed at compile time");
-        sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<T>>(builder);");
-        sb.AppendLine($"        return __b.Select(selector);");
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Where() interceptor with SQL fragment and parameter binder.
-    /// </summary>
-    private static void GenerateWhereInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, List<CachedExtractorField> staticFields, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-
-        // Check if there are captured parameters that need runtime extraction
-        var hasAnyParams = clauseInfo?.Parameters.Count > 0;
-        var hasResolvableCapturedParams = clauseInfo?.Parameters.Any(p => p.IsCaptured && p.CanGenerateDirectPath) == true;
-        var exprParamName = hasResolvableCapturedParams ? "expr" : "_";
-
-        // Emit trim suppression if we'll use FieldInfo.GetValue inline
-        var methodFields = staticFields.Where(f => f.MethodName == methodName).ToList();
-        if (methodFields.Count > 0)
-        {
-            sb.AppendLine($"    [UnconditionalSuppressMessage(\"Trimming\", \"IL2075\",");
-            sb.AppendLine($"        Justification = \"Closure fields are preserved by the expression tree that references them.\")]");
-        }
-
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-
-        // Check if this is on QueryBuilder<T> or QueryBuilder<T, TResult>
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            sb.AppendLine($"    public static {thisType}<{entityType}, {resultType}> {methodName}(");
-            sb.AppendLine($"        this {thisType}<{entityType}, {resultType}> builder,");
-            sb.AppendLine($"        Expression<Func<{entityType}, bool>> {exprParamName})");
-        }
-        else
-        {
-            sb.AppendLine($"    public static {thisType}<{entityType}> {methodName}(");
-            sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-            sb.AppendLine($"        Expression<Func<{entityType}, bool>> {exprParamName})");
-        }
-
-        sb.AppendLine($"    {{");
-
-        if (clauseInfo == null || !clauseInfo.IsSuccess)
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}, {resultType}>>(builder);");
-        }
-        else
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: BindParam instead of AddWhereClause
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-
-            if (hasAnyParams == true)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                var allParams = clauseInfo!.Parameters.OrderBy(p => p.Index).ToList();
-                foreach (var p in allParams)
-                {
-                    var expr = p.IsCaptured ? $"p{p.Index}" : p.ValueExpression;
-                    sb.AppendLine($"        {builderVar}.BindParam({WrapWithToDb(expr, p)});");
-                }
-            }
-
-            if (clauseBit.HasValue)
-                sb.AppendLine($"        return {builderVar}.SetClauseBit({clauseBit.Value});");
-            else
-                sb.AppendLine($"        return {builderVar};");
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        {
-            // Generate optimized interceptor with pre-computed SQL
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-
-            // Check if any captured parameters lack extraction paths (e.g., captured variables in subqueries)
-            var hasUnresolvableCaptured = clauseInfo.Parameters.Any(p => p.IsCaptured && !p.CanGenerateDirectPath);
-
-            var bitSuffix = ClauseBitSuffix(clauseBit);
-            if (hasUnresolvableCaptured)
-            {
-                // Emit SQL-only clause (parameters cannot be extracted at compile time)
-                var resolvableParams = clauseInfo.Parameters
-                    .Where(p => !p.IsCaptured)
-                    .OrderBy(p => p.Index)
-                    .ToList();
-                if (resolvableParams.Count > 0)
-                {
-                    var paramArgs = string.Join(", ", resolvableParams.Select(p => WrapWithToDb(p.ValueExpression, p)));
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\", {paramArgs}){bitSuffix};");
-                }
-                else
-                {
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-                }
-            }
-            else if (hasAnyParams)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                // Build args in index order, mixing captured extractions + literal values
-                var allParams = clauseInfo.Parameters.OrderBy(p => p.Index).ToList();
-                var paramArgs = string.Join(", ", allParams.Select(p =>
-                {
-                    var expr = p.IsCaptured ? $"p{p.Index}" : p.ValueExpression;
-                    return WrapWithToDb(expr, p);
-                }));
-                sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\", {paramArgs}){bitSuffix};");
-            }
-            else
-            {
-                sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-            }
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Where() interceptor for DeleteBuilder or ExecutableDeleteBuilder.
-    /// The return type is always IExecutableDeleteBuilder&lt;T&gt; since Where() on
-    /// DeleteBuilder returns IExecutableDeleteBuilder, and on ExecutableDeleteBuilder returns itself.
-    /// </summary>
-    private static void GenerateDeleteWhereInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, List<CachedExtractorField> staticFields, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-
-        // Check if there are captured parameters that need runtime extraction
-        var hasAnyParams = clauseInfo?.Parameters.Count > 0;
-        var hasResolvableCapturedParams = clauseInfo?.Parameters.Any(p => p.IsCaptured && p.CanGenerateDirectPath) == true;
-        var exprParamName = hasResolvableCapturedParams ? "expr" : "_";
-
-        // Emit trim suppression if we'll use FieldInfo.GetValue inline
-        var methodFields = staticFields.Where(f => f.MethodName == methodName).ToList();
-        if (methodFields.Count > 0)
-        {
-            sb.AppendLine($"    [UnconditionalSuppressMessage(\"Trimming\", \"IL2075\",");
-            sb.AppendLine($"        Justification = \"Closure fields are preserved by the expression tree that references them.\")]");
-        }
-
-        // Determine the receiver type: DeleteBuilder<T> or ExecutableDeleteBuilder<T> (or interface variants)
-        var thisType = site.BuilderTypeName;
-        var isExecutable = thisType.Contains("ExecutableDeleteBuilder");
-        var concreteType = ToConcreteTypeName(thisType);
-        var receiverType = $"{thisType}<{entityType}>";
-
-        sb.AppendLine($"    public static IExecutableDeleteBuilder<{entityType}> {methodName}(");
-        sb.AppendLine($"        this {receiverType} builder,");
-        sb.AppendLine($"        Expression<Func<{entityType}, bool>> {exprParamName})");
-        sb.AppendLine($"    {{");
-
-        if (clauseInfo == null || !clauseInfo.IsSuccess)
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        // Cast to concrete type (receiver is always an interface)
-        sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: BindParam instead of AddWhereClause
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-
-            if (hasAnyParams == true)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                var allParams = clauseInfo!.Parameters.OrderBy(p => p.Index).ToList();
-                foreach (var p in allParams)
-                {
-                    var expr = p.IsCaptured ? $"p{p.Index}" : p.ValueExpression;
-                    sb.AppendLine($"        {builderVar}.BindParam({WrapWithToDb(expr, p)});");
-                }
-            }
-
-            // Non-executable receiver (IDeleteBuilder) must transition to executable via AsExecutable()
-            var returnExpr = isExecutable ? builderVar : $"{builderVar}.AsExecutable()";
-            if (clauseBit.HasValue)
-                sb.AppendLine($"        return {returnExpr}.SetClauseBit({clauseBit.Value});");
-            else
-                sb.AppendLine($"        return {returnExpr};");
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        {
-            // Generate optimized interceptor with pre-computed SQL
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-
-            // Check if any captured parameters lack extraction paths
-            var hasUnresolvableCaptured = clauseInfo.Parameters.Any(p => p.IsCaptured && !p.CanGenerateDirectPath);
-            var bitSuffix = ClauseBitSuffix(clauseBit);
-
-            if (hasUnresolvableCaptured)
-            {
-                // Emit SQL-only clause (parameters cannot be extracted at compile time)
-                var resolvableParams = clauseInfo.Parameters
-                    .Where(p => !p.IsCaptured)
-                    .OrderBy(p => p.Index)
-                    .ToList();
-                if (resolvableParams.Count > 0)
-                {
-                    // Use AddParameter with index capture so @pN placeholders account for
-                    // any prior parameters
-                    var transformedSql = escapedSql;
-                    for (int i = resolvableParams.Count - 1; i >= 0; i--)
-                    {
-                        var p = resolvableParams[i];
-                        sb.AppendLine($"        var _pi{p.Index} = {builderVar}.AddParameter({WrapWithToDb(p.ValueExpression, p)});");
-                        transformedSql = transformedSql.Replace($"@p{p.Index}", $"@p{{_pi{p.Index}}}");
-                    }
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause($@\"{transformedSql}\"){bitSuffix};");
-                }
-                else
-                {
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-                }
-            }
-            else if (hasAnyParams)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                // Add parameters via AddParameter, capturing the runtime index so @pN
-                // placeholders account for any prior parameters
-                var allParams = clauseInfo.Parameters.OrderBy(p => p.Index).ToList();
-                var transformedSql = escapedSql;
-                for (int i = allParams.Count - 1; i >= 0; i--)
-                {
-                    var p = allParams[i];
-                    var valueExpr = p.IsCaptured ? $"p{p.Index}" : p.ValueExpression;
-                    sb.AppendLine($"        var _pi{p.Index} = {builderVar}.AddParameter({WrapWithToDb(valueExpr, p)});");
-                    transformedSql = transformedSql.Replace($"@p{p.Index}", $"@p{{_pi{p.Index}}}");
-                }
-                sb.AppendLine($"        return {builderVar}.AddWhereClause($@\"{transformedSql}\"){bitSuffix};");
-            }
-            else
-            {
-                sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-            }
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates inline expression tree navigation code for a path segment.
-    /// Each segment maps to a specific Unsafe.As cast and property access.
-    /// </summary>
-    private static string GenerateSegmentNavigation(string segment, string prevExpr)
-    {
-        // Handle Arguments[N] format
-        if (segment.StartsWith("Arguments["))
-        {
-            return $"Unsafe.As<MethodCallExpression>({prevExpr}).Arguments[{segment.Substring("Arguments[".Length, 1)}]";
-        }
-
-        return segment switch
-        {
-            "Body" => $"{prevExpr}.Body",
-            "Left" => $"Unsafe.As<BinaryExpression>({prevExpr}).Left",
-            "Right" => $"Unsafe.As<BinaryExpression>({prevExpr}).Right",
-            "Operand" => $"Unsafe.As<UnaryExpression>({prevExpr}).Operand",
-            "Object" => $"Unsafe.As<MethodCallExpression>({prevExpr}).Object!",
-            "Expression" => $"Unsafe.As<MemberExpression>({prevExpr}).Expression!",
-            "Test" => $"Unsafe.As<ConditionalExpression>({prevExpr}).Test",
-            "IfTrue" => $"Unsafe.As<ConditionalExpression>({prevExpr}).IfTrue",
-            "IfFalse" => $"Unsafe.As<ConditionalExpression>({prevExpr}).IfFalse",
-            _ => throw new InvalidOperationException($"Unknown path segment: {segment}")
-        };
-    }
-
-    /// <summary>
-    /// Generates inline navigation code that arrives at the MemberExpression for a given path.
-    /// Returns the variable name holding the final MemberExpression.
-    /// </summary>
-    /// <remarks>
-    /// For enum captured variables, the C# compiler inserts a UnaryExpression(Convert) node
-    /// wrapping the MemberExpression. The generated code uses a pattern match to unwrap it.
-    /// </remarks>
-    private static string GenerateInlineNavigation(StringBuilder sb, string path, int paramIndex)
-    {
-        var segments = path.Split('.');
-        // "Body" is the first segment — navigate from expr
-        // The remaining segments (except the last, which is the MemberExpression itself)
-        // are intermediate casts. We build a single chained expression.
-
-        // Build the chain from the inside out
-        var currentExpr = "expr";
-        foreach (var segment in segments)
-        {
-            currentExpr = GenerateSegmentNavigation(segment, currentExpr);
-        }
-
-        // The final node is a MemberExpression, but for enum captured variables
-        // the compiler inserts a UnaryExpression(Convert) wrapper that must be stripped.
-        var nodeVar = $"_n{paramIndex}";
-        var memberVar = $"_m{paramIndex}";
-        sb.AppendLine($"        var {nodeVar} = {currentExpr};");
-        sb.AppendLine($"        var {memberVar} = {nodeVar} is UnaryExpression _u{paramIndex} ? Unsafe.As<MemberExpression>(_u{paramIndex}.Operand) : Unsafe.As<MemberExpression>({nodeVar});");
-        return memberVar;
-    }
-
-    /// <summary>
-    /// Generates cached extractor code using static fields for all captured parameters.
-    /// Emits inline Unsafe.As navigation with cached FieldInfo.
-    /// </summary>
-    private static void GenerateCachedExtraction(StringBuilder sb, List<CachedExtractorField> fields)
-    {
-        foreach (var field in fields)
-        {
-            sb.AppendLine($"        // Inline extraction: {field.ExpressionPath}");
-            var memberVar = GenerateInlineNavigation(sb, field.ExpressionPath, field.ParameterIndex);
-            sb.AppendLine($"        {field.FieldName} ??= Unsafe.As<FieldInfo>({memberVar}.Member);");
-            sb.AppendLine($"        var p{field.ParameterIndex} = {field.FieldName}.GetValue(Unsafe.As<ConstantExpression>({memberVar}.Expression!).Value);");
-        }
-    }
-
-    /// <summary>
-    /// Generates an OrderBy/ThenBy interceptor with SQL fragment.
-    /// </summary>
-    private static void GenerateOrderByInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-        var isOrderBy = site.Kind == InterceptorKind.OrderBy;
-
-        // Use concrete key type (arity 0) when available; otherwise fall back to arity-matching
-        var keyType = site.KeyTypeName != null ? GetShortTypeName(site.KeyTypeName) : null;
-
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-
-        // Check if this is on QueryBuilder<T> or QueryBuilder<T, TResult>
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            if (keyType != null)
-            {
-                // Non-generic interceptor (arity 0) — concrete key type
-                sb.AppendLine($"    public static {thisType}<{entityType}, {resultType}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}, {resultType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, {keyType}>> _,");
-                sb.AppendLine($"        Direction direction = Direction.Ascending)");
-            }
-            else
-            {
-                // Arity-matching interceptor — include all type params from class + method
-                sb.AppendLine($"    public static {thisType}<T, TResult> {methodName}<T, TResult, TKey>(");
-                sb.AppendLine($"        this {thisType}<T, TResult> builder,");
-                sb.AppendLine($"        Expression<Func<T, TKey>> _,");
-                sb.AppendLine($"        Direction direction = Direction.Ascending) where T : class");
-            }
-        }
-        else
-        {
-            if (keyType != null)
-            {
-                // Non-generic interceptor (arity 0) — concrete key type
-                sb.AppendLine($"    public static {thisType}<{entityType}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, {keyType}>> _,");
-                sb.AppendLine($"        Direction direction = Direction.Ascending)");
-            }
-            else
-            {
-                // Arity-matching interceptor — include all type params from class + method
-                sb.AppendLine($"    public static {thisType}<T> {methodName}<T, TKey>(");
-                sb.AppendLine($"        this {thisType}<T> builder,");
-                sb.AppendLine($"        Expression<Func<T, TKey>> _,");
-                sb.AppendLine($"        Direction direction = Direction.Ascending) where T : class");
-            }
-        }
-
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}, {resultType}>>(builder);");
-        }
-        else if (keyType != null)
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-        }
-        else
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<T>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: passthrough (no SQL generation)
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            if (clauseBit.HasValue)
-            {
-                sb.AppendLine($"        {builderVar}.SetClauseBit({clauseBit.Value});");
-                sb.AppendLine($"        return builder;");
-            }
-            else
-            {
-                sb.AppendLine($"        return builder;");
-            }
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        // When using open generic type params (arity-matching), the concrete return type
-        // can't be implicitly converted to the open generic interface return type.
-        // Cast through Unsafe.As to bridge the gap.
-        var needsReturnCast = keyType == null;
-        string returnCastOpen;
-        string returnCastClose;
-        if (needsReturnCast)
-        {
-            var returnType = site.ResultTypeName != null ? $"{thisType}<T, TResult>" : $"{thisType}<T>";
-            returnCastOpen = $"Unsafe.As<{returnType}>(";
-            returnCastClose = ")";
-        }
-        else
-        {
-            returnCastOpen = "";
-            returnCastClose = "";
-        }
-
-        var bitSuffix = ClauseBitSuffix(clauseBit);
-        if (clauseInfo is OrderByClauseInfo orderByInfo && orderByInfo.IsSuccess)
-        {
-            // Generate optimized interceptor with pre-computed SQL
-            var escapedSql = EscapeStringLiteral(orderByInfo.ColumnSql);
-            var directionSuffix = orderByInfo.IsDescending ? " DESC" : " ASC";
-            var builderMethod = isOrderBy ? "AddOrderByClause" : "AddThenByClause";
-            sb.AppendLine($"        return {returnCastOpen}{builderVar}.{builderMethod}(@\"{escapedSql}\", direction){bitSuffix}{returnCastClose};");
-        }
-        else if (clauseInfo != null && clauseInfo.IsSuccess)
-        {
-            // Non-OrderByClauseInfo but still successful
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-            var builderMethod = isOrderBy ? "AddOrderByClause" : "AddThenByClause";
-            sb.AppendLine($"        return {returnCastOpen}{builderVar}.{builderMethod}(@\"{escapedSql}\", direction){bitSuffix}{returnCastClose};");
-        }
-        else
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a GroupBy interceptor with SQL fragment.
-    /// </summary>
-    private static void GenerateGroupByInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-
-        // Use concrete key type (arity 0) when available; otherwise fall back to arity-matching
-        var keyType = site.KeyTypeName != null ? GetShortTypeName(site.KeyTypeName) : null;
-
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-
-        // Check if this is on QueryBuilder<T> or QueryBuilder<T, TResult>
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            if (keyType != null)
-            {
-                sb.AppendLine($"    public static {thisType}<{entityType}, {resultType}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}, {resultType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, {keyType}>> _)");
-            }
-            else
-            {
-                sb.AppendLine($"    public static {thisType}<T, TResult> {methodName}<T, TResult, TKey>(");
-                sb.AppendLine($"        this {thisType}<T, TResult> builder,");
-                sb.AppendLine($"        Expression<Func<T, TKey>> _) where T : class");
-            }
-        }
-        else
-        {
-            if (keyType != null)
-            {
-                sb.AppendLine($"    public static {thisType}<{entityType}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, {keyType}>> _)");
-            }
-            else
-            {
-                sb.AppendLine($"    public static {thisType}<T> {methodName}<T, TKey>(");
-                sb.AppendLine($"        this {thisType}<T> builder,");
-                sb.AppendLine($"        Expression<Func<T, TKey>> _) where T : class");
-            }
-        }
-
-        sb.AppendLine($"    {{");
-
-        if (clauseInfo == null || !clauseInfo.IsSuccess)
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}, {resultType}>>(builder);");
-        }
-        else if (keyType != null)
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-        }
-        else
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<T>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: passthrough (no SQL generation)
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            if (clauseBit.HasValue)
-            {
-                sb.AppendLine($"        {builderVar}.SetClauseBit({clauseBit.Value});");
-                sb.AppendLine($"        return builder;");
-            }
-            else
-            {
-                sb.AppendLine($"        return builder;");
-            }
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        {
-            // Generate optimized interceptor with pre-computed SQL
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-            var bitSuffix = ClauseBitSuffix(clauseBit);
-
-            // When using open generic type params (arity-matching), cast return through Unsafe.As
-            var needsReturnCast = keyType == null;
-            string returnCastOpen = "", returnCastClose = "";
-            if (needsReturnCast)
-            {
-                var returnType = site.ResultTypeName != null ? $"{thisType}<T, TResult>" : $"{thisType}<T>";
-                returnCastOpen = $"Unsafe.As<{returnType}>(";
-                returnCastClose = ")";
-            }
-
-            sb.AppendLine($"        return {returnCastOpen}{builderVar}.AddGroupByClause(@\"{escapedSql}\"){bitSuffix}{returnCastClose};");
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Having interceptor with SQL fragment.
-    /// </summary>
-    private static void GenerateHavingInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-
-        // Check if this is on QueryBuilder<T> or QueryBuilder<T, TResult>
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            sb.AppendLine($"    public static {thisType}<{entityType}, {resultType}> {methodName}(");
-            sb.AppendLine($"        this {thisType}<{entityType}, {resultType}> builder,");
-            sb.AppendLine($"        Expression<Func<{entityType}, bool>> _)");
-        }
-        else
-        {
-            sb.AppendLine($"    public static {thisType}<{entityType}> {methodName}(");
-            sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-            sb.AppendLine($"        Expression<Func<{entityType}, bool>> _)");
-        }
-
-        sb.AppendLine($"    {{");
-
-        if (clauseInfo == null || !clauseInfo.IsSuccess)
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}, {resultType}>>(builder);");
-        }
-        else
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: passthrough (no SQL generation)
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            if (clauseBit.HasValue)
-            {
-                sb.AppendLine($"        {builderVar}.SetClauseBit({clauseBit.Value});");
-                sb.AppendLine($"        return builder;");
-            }
-            else
-            {
-                sb.AppendLine($"        return builder;");
-            }
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        {
-            // Generate optimized interceptor with pre-computed SQL
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-            var bitSuffix = ClauseBitSuffix(clauseBit);
-            sb.AppendLine($"        return {builderVar}.AddHavingClause(@\"{escapedSql}\"){bitSuffix};");
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Set() interceptor with SQL fragment.
-    /// </summary>
-    private static void GenerateSetInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-
-        sb.AppendLine($"    public static {thisType}<{entityType}> {methodName}<TValue>(");
-        sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-        sb.AppendLine($"        Expression<Func<{entityType}, TValue>> _,");
-        sb.AppendLine($"        TValue value)");
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: BindParam for value
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            var setValueArg = (clauseInfo is SetClauseInfo prebuiltSetInfo && prebuiltSetInfo.CustomTypeMappingClass != null)
-                ? $"{GetMappingFieldName(prebuiltSetInfo.CustomTypeMappingClass)}.ToDb(value)"
-                : "value";
-            sb.AppendLine($"        {builderVar}.BindParam({setValueArg});");
-            if (clauseBit.HasValue)
-                sb.AppendLine($"        return {builderVar}.SetClauseBit({clauseBit.Value});");
-            else
-                sb.AppendLine($"        return {builderVar};");
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        if (clauseInfo is SetClauseInfo setInfo && setInfo.IsSuccess)
-        {
-            // Generate optimized interceptor with pre-computed column
-            var escapedColumnSql = EscapeStringLiteral(setInfo.ColumnSql);
-            var valueArg = setInfo.CustomTypeMappingClass != null
-                ? $"{GetMappingFieldName(setInfo.CustomTypeMappingClass)}.ToDb(value)"
-                : "value";
-            var bitSuffix = ClauseBitSuffix(clauseBit);
-            sb.AppendLine($"        return {builderVar}.AddSetClause(@\"{escapedColumnSql}\", {valueArg}, {setInfo.ParameterIndex}){bitSuffix};");
-        }
-        else if (clauseInfo != null && clauseInfo.IsSuccess)
-        {
-            // Non-SetClauseInfo but still successful
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-            var bitSuffix = ClauseBitSuffix(clauseBit);
-            sb.AppendLine($"        return {builderVar}.AddSetClauseRaw(@\"{escapedSql}\", value){bitSuffix};");
-        }
-        else
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Set() interceptor for UpdateBuilder or ExecutableUpdateBuilder.
-    /// The return type uses the interface variant: IUpdateBuilder&lt;T&gt; or IExecutableUpdateBuilder&lt;T&gt;.
-    /// </summary>
-    private static void GenerateUpdateSetInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-
-        // Determine the receiver type: UpdateBuilder<T> or ExecutableUpdateBuilder<T> (or interface variants)
-        var thisType = site.BuilderTypeName;
-        var isExecutable = thisType.Contains("ExecutableUpdateBuilder");
-        var concreteBaseName = isExecutable ? "ExecutableUpdateBuilder" : "UpdateBuilder";
-        var returnInterfaceBaseName = "I" + concreteBaseName;
-        // Interceptors for generic methods on generic types need arity = type params + method params.
-        // UpdateBuilder<T>.Set<TValue>() has total arity 2 (T + TValue).
-        // We emit the interceptor with <T, TValue> and where T : class constraint.
-        sb.AppendLine($"    public static {returnInterfaceBaseName}<T> {methodName}<T, TValue>(");
-        sb.AppendLine($"        this {thisType}<T> builder,");
-        sb.AppendLine($"        Expression<Func<T, TValue>> _,");
-        sb.AppendLine($"        TValue value) where T : class");
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        sb.AppendLine($"        var __b = Unsafe.As<{concreteBaseName}<T>>(builder);");
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: BindParam for value
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            sb.AppendLine($"        {builderVar}.BindParam(value);");
-            if (clauseBit.HasValue)
-                sb.AppendLine($"        return {builderVar}.SetClauseBit({clauseBit.Value});");
-            else
-                sb.AppendLine($"        return {builderVar};");
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        var bitSuffix = ClauseBitSuffix(clauseBit);
-        if (clauseInfo is SetClauseInfo setInfo && setInfo.IsSuccess)
-        {
-            // Generate optimized interceptor with pre-computed column (from semantic analysis)
-            var escapedColumnSql = EscapeStringLiteral(setInfo.ColumnSql);
-            sb.AppendLine($"        return {builderVar}.AddSetClause(@\"{escapedColumnSql}\", value){bitSuffix};");
-        }
-        else if (clauseInfo != null && clauseInfo.IsSuccess)
-        {
-            // Plain ClauseInfo from syntactic translation — SqlFragment is the column SQL
-            var escapedColumnSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-            sb.AppendLine($"        return {builderVar}.AddSetClause(@\"{escapedColumnSql}\", value){bitSuffix};");
-        }
-        else
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Set(T entity) POCO interceptor for UpdateBuilder or ExecutableUpdateBuilder.
-    /// Extracts property values from the entity and calls AddSetClause for each initialized column.
-    /// </summary>
-    private static void GenerateUpdateSetPocoInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var updateInfo = site.UpdateInfo;
-
-        // Determine the receiver type: UpdateBuilder<T> or ExecutableUpdateBuilder<T> (or interface variants)
-        var thisType = site.BuilderTypeName;
-        var isExecutable = thisType.Contains("ExecutableUpdateBuilder");
-        var concreteBaseName = isExecutable ? "ExecutableUpdateBuilder" : "UpdateBuilder";
-        var returnInterfaceBaseName = "I" + concreteBaseName;
-        // Set(T entity) has no method-level type params, only class-level T.
-        // Interceptor arity = 1 (T from the class).
-        sb.AppendLine($"    public static {returnInterfaceBaseName}<T> {methodName}<T>(");
-        sb.AppendLine($"        this {thisType}<T> builder,");
-        sb.AppendLine($"        T entity) where T : class");
-        sb.AppendLine($"    {{");
-
-        if (updateInfo != null && updateInfo.Columns.Count > 0)
-        {
-            // Cast to concrete type (receiver is always an interface)
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBaseName}<T>>(builder);");
-            var builderVar = "__b";
-
-            // Cast to concrete entity type for property access
-            sb.AppendLine($"        var e = Unsafe.As<{entityType}>(entity);");
-
-            foreach (var column in updateInfo.Columns)
-            {
-                var escapedColumnSql = EscapeStringLiteral(column.QuotedColumnName);
-                var valueExpr = column.IsForeignKey
-                    ? $"e.{column.PropertyName}.Id"
-                    : $"e.{column.PropertyName}";
-                if (column.CustomTypeMappingClass != null)
-                    valueExpr = $"{GetMappingFieldName(column.CustomTypeMappingClass)}.ToDb({valueExpr})";
-
-                var sensitiveArg = column.IsSensitive ? ", isSensitive: true" : "";
-                sb.AppendLine($"        {builderVar}.AddSetClause(@\"{escapedColumnSql}\", {valueExpr}{sensitiveArg});");
-            }
-
-            sb.AppendLine($"        return {builderVar};");
-        }
-        else
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Where() interceptor for UpdateBuilder or ExecutableUpdateBuilder.
-    /// The return type is always IExecutableUpdateBuilder&lt;T&gt; since Where() on
-    /// UpdateBuilder returns IExecutableUpdateBuilder, and on ExecutableUpdateBuilder returns itself.
-    /// </summary>
-    private static void GenerateUpdateWhereInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, List<CachedExtractorField> staticFields, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo;
-
-        // Check if there are captured parameters that need runtime extraction
-        var hasAnyParams = clauseInfo?.Parameters.Count > 0;
-        var hasResolvableCapturedParams = clauseInfo?.Parameters.Any(p => p.IsCaptured && p.CanGenerateDirectPath) == true;
-        var exprParamName = hasResolvableCapturedParams ? "expr" : "_";
-
-        // Emit trim suppression if we'll use FieldInfo.GetValue inline
-        var methodFields = staticFields.Where(f => f.MethodName == methodName).ToList();
-        if (methodFields.Count > 0)
-        {
-            sb.AppendLine($"    [UnconditionalSuppressMessage(\"Trimming\", \"IL2075\",");
-            sb.AppendLine($"        Justification = \"Closure fields are preserved by the expression tree that references them.\")]");
-        }
-
-        // Determine the receiver type: UpdateBuilder<T> or ExecutableUpdateBuilder<T> (or interface variants)
-        var thisType = site.BuilderTypeName;
-        var isExecutable = thisType.Contains("ExecutableUpdateBuilder");
-        var concreteType = ToConcreteTypeName(thisType);
-        var receiverType = $"{thisType}<{entityType}>";
-
-        sb.AppendLine($"    public static IExecutableUpdateBuilder<{entityType}> {methodName}(");
-        sb.AppendLine($"        this {receiverType} builder,");
-        sb.AppendLine($"        Expression<Func<{entityType}, bool>> {exprParamName})");
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}>>(builder);");
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: BindParam instead of AddWhereClause
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-
-            if (hasAnyParams == true)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                var allParams = clauseInfo!.Parameters.OrderBy(p => p.Index).ToList();
-                foreach (var p in allParams)
-                {
-                    var expr = p.IsCaptured ? $"p{p.Index}" : p.ValueExpression;
-                    sb.AppendLine($"        {builderVar}.BindParam({WrapWithToDb(expr, p)});");
-                }
-            }
-
-            // Non-executable receiver (IUpdateBuilder) must transition to executable via AsExecutable()
-            var returnExpr = isExecutable ? builderVar : $"{builderVar}.AsExecutable()";
-            if (clauseBit.HasValue)
-                sb.AppendLine($"        return {returnExpr}.SetClauseBit({clauseBit.Value});");
-            else
-                sb.AppendLine($"        return {returnExpr};");
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        if (clauseInfo != null && clauseInfo.IsSuccess)
-        {
-            // Generate optimized interceptor with pre-computed SQL
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-
-            // Check if any captured parameters lack extraction paths
-            var hasUnresolvableCaptured = clauseInfo.Parameters.Any(p => p.IsCaptured && !p.CanGenerateDirectPath);
-            var bitSuffix = ClauseBitSuffix(clauseBit);
-
-            if (hasUnresolvableCaptured)
-            {
-                // Emit SQL-only clause (parameters cannot be extracted at compile time)
-                var resolvableParams = clauseInfo.Parameters
-                    .Where(p => !p.IsCaptured)
-                    .OrderBy(p => p.Index)
-                    .ToList();
-                if (resolvableParams.Count > 0)
-                {
-                    // Use AddParameter with index capture so @pN placeholders account for
-                    // any prior parameters (e.g. from SET clauses)
-                    var transformedSql = escapedSql;
-                    for (int i = resolvableParams.Count - 1; i >= 0; i--)
-                    {
-                        var p = resolvableParams[i];
-                        sb.AppendLine($"        var _pi{p.Index} = {builderVar}.AddParameter({WrapWithToDb(p.ValueExpression, p)});");
-                        transformedSql = transformedSql.Replace($"@p{p.Index}", $"@p{{_pi{p.Index}}}");
-                    }
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause($@\"{transformedSql}\"){bitSuffix};");
-                }
-                else
-                {
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-                }
-            }
-            else if (hasAnyParams)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                // Add parameters via AddParameter, capturing the runtime index so @pN
-                // placeholders account for any prior parameters (e.g. from SET clauses)
-                var allParams = clauseInfo.Parameters.OrderBy(p => p.Index).ToList();
-                var transformedSql = escapedSql;
-                for (int i = allParams.Count - 1; i >= 0; i--)
-                {
-                    var p = allParams[i];
-                    var valueExpr = p.IsCaptured ? $"p{p.Index}" : p.ValueExpression;
-                    sb.AppendLine($"        var _pi{p.Index} = {builderVar}.AddParameter({WrapWithToDb(valueExpr, p)});");
-                    transformedSql = transformedSql.Replace($"@p{p.Index}", $"@p{{_pi{p.Index}}}");
-                }
-                sb.AppendLine($"        return {builderVar}.AddWhereClause($@\"{transformedSql}\"){bitSuffix};");
-            }
-            else
-            {
-                sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-            }
-        }
-        else
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Join interceptor with SQL fragment generation.
-    /// Handles both initial joins (QueryBuilder→JoinedQueryBuilder) and
-    /// chained joins (JoinedQueryBuilder→JoinedQueryBuilder3, etc.)
-    /// </summary>
-    private static void GenerateJoinInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var clauseInfo = site.ClauseInfo as JoinClauseInfo;
-        var joinedEntityTypeNames = site.JoinedEntityTypeNames;
-
-        var joinKind = site.Kind switch
-        {
-            InterceptorKind.LeftJoin => "JoinKind.Left",
-            InterceptorKind.RightJoin => "JoinKind.Right",
-            _ => "JoinKind.Inner"
-        };
-
-        var thisType = site.BuilderTypeName;
-        var concreteThisType = ToConcreteTypeName(thisType);
-
-        // Determine if this is a chained join (from JoinedQueryBuilder/3)
-        var isChainedJoin = joinedEntityTypeNames != null && joinedEntityTypeNames.Count >= 2;
-
-        if (isChainedJoin && site.JoinedEntityTypeName != null)
-        {
-            var joinedType = GetShortTypeName(site.JoinedEntityTypeName);
-            var priorTypes = joinedEntityTypeNames!.Select(GetShortTypeName).ToList();
-            var allTypes = priorTypes.Concat(new[] { joinedType }).ToList();
-
-            // Determine receiver and return builder type names
-            var receiverBuilderName = GetJoinedBuilderTypeName(priorTypes.Count);
-            var concreteReceiverBuilderName = ToConcreteTypeName(receiverBuilderName);
-            var returnBuilderName = GetJoinedBuilderTypeName(allTypes.Count);
-            var receiverTypeArgs = string.Join(", ", priorTypes);
-            var returnTypeArgs = string.Join(", ", allTypes);
-            var funcTypeArgs = string.Join(", ", allTypes) + ", bool";
-
-            if (clauseInfo != null && clauseInfo.IsSuccess)
-            {
-                var escapedSql = EscapeStringLiteral(clauseInfo.OnConditionSql);
-                sb.AppendLine($"    public static {returnBuilderName}<{returnTypeArgs}> {methodName}(");
-                sb.AppendLine($"        this {receiverBuilderName}<{receiverTypeArgs}> builder,");
-                sb.AppendLine($"        Expression<Func<{funcTypeArgs}>> _)");
-                sb.AppendLine($"    {{");
-                sb.AppendLine($"        var __b = Unsafe.As<{concreteReceiverBuilderName}<{receiverTypeArgs}>>(builder);");
-                var escapedTableName = EscapeStringLiteral(clauseInfo.JoinedTableName);
-                sb.AppendLine($"        return __b.AddJoinClause<{joinedType}>({joinKind}, @\"{escapedTableName}\", @\"{escapedSql}\");");
-                sb.AppendLine($"    }}");
-            }
-            else
-            {
-                var methodCall = site.Kind switch
-                {
-                    InterceptorKind.LeftJoin => "LeftJoin",
-                    InterceptorKind.RightJoin => "RightJoin",
-                    _ => "Join"
-                };
-
-                sb.AppendLine($"    public static {returnBuilderName}<{returnTypeArgs}> {methodName}(");
-                sb.AppendLine($"        this {receiverBuilderName}<{receiverTypeArgs}> builder,");
-                sb.AppendLine($"        Expression<Func<{funcTypeArgs}>> condition)");
-                sb.AppendLine($"    {{");
-                sb.AppendLine($"        var __b = Unsafe.As<{concreteReceiverBuilderName}<{receiverTypeArgs}>>(builder);");
-                sb.AppendLine($"        // Fallback - join condition not fully analyzed at compile time");
-                sb.AppendLine($"        return __b.{methodCall}(condition);");
-                sb.AppendLine($"    }}");
-            }
-        }
-        else if (clauseInfo != null && clauseInfo.IsSuccess)
-        {
-            var joinedEntityName = clauseInfo.JoinedEntityName;
-            var escapedSql = EscapeStringLiteral(clauseInfo.OnConditionSql);
-
-            if (site.IsNavigationJoin)
-            {
-                // Navigation overload: Expression<Func<T, NavigationList<TJoined>>>
-                sb.AppendLine($"    public static IJoinedQueryBuilder<{entityType}, {joinedEntityName}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, NavigationList<{joinedEntityName}>>> _)");
-            }
-            else
-            {
-                // Explicit-lambda overload: Expression<Func<T, TJoined, bool>>
-                sb.AppendLine($"    public static IJoinedQueryBuilder<{entityType}, {joinedEntityName}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, {joinedEntityName}, bool>> _)");
-            }
-            sb.AppendLine($"    {{");
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteThisType}<{entityType}>>(builder);");
-            var escapedTableName = EscapeStringLiteral(clauseInfo.JoinedTableName);
-            sb.AppendLine($"        return __b.AddJoinClause<{joinedEntityName}>({joinKind}, @\"{escapedTableName}\", @\"{escapedSql}\");");
-            sb.AppendLine($"    }}");
-        }
-        else if (site.JoinedEntityTypeName != null)
-        {
-            // Fallback with concrete joined type - delegates to original method
-            var joinedType = GetShortTypeName(site.JoinedEntityTypeName);
-
-            var methodCall = site.Kind switch
-            {
-                InterceptorKind.LeftJoin => "LeftJoin",
-                InterceptorKind.RightJoin => "RightJoin",
-                _ => "Join"
-            };
-
-            if (site.IsNavigationJoin)
-            {
-                // Navigation overload fallback
-                sb.AppendLine($"    public static IJoinedQueryBuilder<{entityType}, {joinedType}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, NavigationList<{joinedType}>>> navigation)");
-                sb.AppendLine($"    {{");
-                sb.AppendLine($"        var __b = Unsafe.As<{concreteThisType}<{entityType}>>(builder);");
-                sb.AppendLine($"        // Fallback - navigation join not fully analyzed at compile time");
-                sb.AppendLine($"        return __b.{methodCall}(navigation);");
-                sb.AppendLine($"    }}");
-            }
-            else
-            {
-                // Explicit-lambda overload fallback
-                sb.AppendLine($"    public static IJoinedQueryBuilder<{entityType}, {joinedType}> {methodName}(");
-                sb.AppendLine($"        this {thisType}<{entityType}> builder,");
-                sb.AppendLine($"        Expression<Func<{entityType}, {joinedType}, bool>> condition)");
-                sb.AppendLine($"    {{");
-                sb.AppendLine($"        var __b = Unsafe.As<{concreteThisType}<{entityType}>>(builder);");
-                sb.AppendLine($"        // Fallback - join condition not fully analyzed at compile time");
-                sb.AppendLine($"        return __b.{methodCall}(condition);");
-                sb.AppendLine($"    }}");
-            }
-        }
-        else
-        {
-            // No type info available - skip this interceptor
-            sb.AppendLine($"    // WARNING: Could not determine joined entity type for {site.MethodName}");
-        }
-    }
-
-    /// <summary>
-    /// Gets the builder type name for a given entity count in joins.
-    /// </summary>
-    private static string GetJoinedBuilderTypeName(int entityCount)
-    {
-        return entityCount switch
-        {
-            2 => "IJoinedQueryBuilder",
-            3 => "IJoinedQueryBuilder3",
-            4 => "IJoinedQueryBuilder4",
-            _ => throw new ArgumentOutOfRangeException(nameof(entityCount), $"Unsupported entity count: {entityCount}")
-        };
-    }
-
-    /// <summary>
-    /// Generates a Where() interceptor for joined query builders.
-    /// </summary>
-    private static void GenerateJoinedWhereInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, List<CachedExtractorField> staticFields, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityTypes = site.JoinedEntityTypeNames!.Select(GetShortTypeName).ToList();
-        var builderName = GetJoinedBuilderTypeName(entityTypes.Count);
-        var concreteBuilderName = ToConcreteTypeName(builderName);
-        var thisType = site.BuilderTypeName;
-        var thisBuilderName = builderName;
-        var typeArgs = string.Join(", ", entityTypes);
-        var clauseInfo = site.ClauseInfo;
-        var hasAnyParams = clauseInfo?.Parameters.Count > 0;
-        var hasResolvableCapturedParams = clauseInfo?.Parameters.Any(p => p.IsCaptured && p.CanGenerateDirectPath) == true;
-        var exprParamName = hasResolvableCapturedParams ? "expr" : "_";
-        var bitSuffix = ClauseBitSuffix(clauseBit);
-
-        var methodFields = staticFields.Where(f => f.MethodName == methodName).ToList();
-        if (methodFields.Count > 0)
-        {
-            sb.AppendLine($"    [UnconditionalSuppressMessage(\"Trimming\", \"IL2075\",");
-            sb.AppendLine($"        Justification = \"Closure fields are preserved by the expression tree that references them.\")]");
-        }
-
-        // Check if this is on a projected builder (has TResult)
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            var isBrokenTuple = resultType.Contains("object") && resultType.StartsWith("(");
-
-            if (isBrokenTuple)
-            {
-                // Tuple element types could not be resolved by the semantic model (generated entity types).
-                // Use arity-matching generic parameters so the compiler infers the concrete TResult.
-                var allTypeParams = string.Join(", ", Enumerable.Range(1, entityTypes.Count).Select(i => $"T{i}"));
-                var constraints = string.Join(" ", Enumerable.Range(1, entityTypes.Count).Select(i => $"where T{i} : class"));
-                sb.AppendLine($"    public static {builderName}<{allTypeParams}, TResult> {methodName}<{allTypeParams}, TResult>(");
-                sb.AppendLine($"        this {thisBuilderName}<{allTypeParams}, TResult> builder,");
-                sb.AppendLine($"        Expression<Func<{allTypeParams}, bool>> {exprParamName}) {constraints}");
-            }
-            else
-            {
-                sb.AppendLine($"    public static {builderName}<{typeArgs}, {resultType}> {methodName}(");
-                sb.AppendLine($"        this {thisBuilderName}<{typeArgs}, {resultType}> builder,");
-                sb.AppendLine($"        Expression<Func<{typeArgs}, bool>> {exprParamName})");
-            }
-        }
-        else
-        {
-            sb.AppendLine($"    public static {builderName}<{typeArgs}> {methodName}(");
-            sb.AppendLine($"        this {thisBuilderName}<{typeArgs}> builder,");
-            sb.AppendLine($"        Expression<Func<{typeArgs}, bool>> {exprParamName})");
-        }
-
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            var isBrokenTuple = resultType.Contains("object") && resultType.StartsWith("(");
-            if (isBrokenTuple)
-            {
-                var allTypeParams = string.Join(", ", Enumerable.Range(1, entityTypes.Count).Select(i => $"T{i}"));
-                sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{allTypeParams}, TResult>>(builder);");
-            }
-            else
-            {
-                sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{typeArgs}, {resultType}>>(builder);");
-            }
-        }
-        else
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{typeArgs}>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: BindParam instead of AddWhereClause
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-
-            if (hasAnyParams == true)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                var allParams = clauseInfo!.Parameters.OrderBy(p => p.Index).ToList();
-                foreach (var p in allParams)
-                {
-                    var expr = p.IsCaptured ? $"p{p.Index}" : p.ValueExpression;
-                    sb.AppendLine($"        {builderVar}.BindParam({expr});");
-                }
-            }
-
-            if (clauseBit.HasValue)
-            {
-                sb.AppendLine($"        {builderVar}.SetClauseBit({clauseBit.Value});");
-                sb.AppendLine($"        return builder;");
-            }
-            else
-            {
-                sb.AppendLine($"        return builder;");
-            }
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        if (clauseInfo != null && clauseInfo.IsSuccess)
-        {
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-
-            // Check if any captured parameters lack extraction paths (e.g., captured variables in subqueries)
-            var hasUnresolvableCaptured = clauseInfo.Parameters.Any(p => p.IsCaptured && !p.CanGenerateDirectPath);
-
-            if (hasUnresolvableCaptured)
-            {
-                // Emit SQL-only clause (parameters cannot be extracted at compile time)
-                var resolvableParams = clauseInfo.Parameters
-                    .Where(p => !p.IsCaptured)
-                    .OrderBy(p => p.Index)
-                    .ToList();
-                if (resolvableParams.Count > 0)
-                {
-                    var paramArgs = string.Join(", ", resolvableParams.Select(p => p.ValueExpression));
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\", {paramArgs}){bitSuffix};");
-                }
-                else
-                {
-                    sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-                }
-            }
-            else if (hasAnyParams)
-            {
-                if (hasResolvableCapturedParams)
-                    GenerateCachedExtraction(sb, methodFields);
-
-                var allParams = clauseInfo.Parameters.OrderBy(p => p.Index).ToList();
-                var paramArgs = string.Join(", ", allParams.Select(p =>
-                    p.IsCaptured ? $"p{p.Index}" : p.ValueExpression));
-                sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\", {paramArgs}){bitSuffix};");
-            }
-            else
-            {
-                sb.AppendLine($"        return {builderVar}.AddWhereClause(@\"{escapedSql}\"){bitSuffix};");
-            }
-        }
-        else
-        {
-            sb.AppendLine($"        // Fallback - expression not fully analyzed at compile time");
-            sb.AppendLine($"        return {builderVar}.Where({exprParamName}){bitSuffix};");
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates an OrderBy/ThenBy interceptor for joined query builders.
-    /// </summary>
-    private static void GenerateJoinedOrderByInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName, int? clauseBit = null,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityTypes = site.JoinedEntityTypeNames!.Select(GetShortTypeName).ToList();
-        var builderName = GetJoinedBuilderTypeName(entityTypes.Count);
-        var concreteBuilderName = ToConcreteTypeName(builderName);
-        var thisType = site.BuilderTypeName;
-        var thisBuilderName = builderName;
-        var typeArgs = string.Join(", ", entityTypes);
-        var clauseInfo = site.ClauseInfo;
-        var isOrderBy = site.Kind == InterceptorKind.OrderBy;
-        var bitSuffix = ClauseBitSuffix(clauseBit);
-
-        // Use concrete key type (arity 0) when available
-        var keyType = site.KeyTypeName != null ? GetShortTypeName(site.KeyTypeName) : null;
-
-        if (site.ResultTypeName != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            var isBrokenTuple = resultType.Contains("object") && resultType.StartsWith("(");
-
-            if (keyType != null && !isBrokenTuple)
-            {
-                sb.AppendLine($"    public static {builderName}<{typeArgs}, {resultType}> {methodName}(");
-                sb.AppendLine($"        this {thisBuilderName}<{typeArgs}, {resultType}> builder,");
-                sb.AppendLine($"        Expression<Func<{typeArgs}, {keyType}>> _,");
-                sb.AppendLine($"        Direction direction = Direction.Ascending)");
-            }
-            else
-            {
-                // Arity-matching: include all type params with class constraints.
-                // Also used when tuple result type has unresolved element types (broken tuple).
-                var allTypeParams = string.Join(", ", Enumerable.Range(1, entityTypes.Count).Select(i => $"T{i}"));
-                var constraints = string.Join(" ", Enumerable.Range(1, entityTypes.Count).Select(i => $"where T{i} : class"));
-                if (keyType != null)
-                {
-                    sb.AppendLine($"    public static {builderName}<{allTypeParams}, TResult> {methodName}<{allTypeParams}, TResult>(");
-                    sb.AppendLine($"        this {thisBuilderName}<{allTypeParams}, TResult> builder,");
-                    sb.AppendLine($"        Expression<Func<{allTypeParams}, {keyType}>> _,");
-                    sb.AppendLine($"        Direction direction = Direction.Ascending) {constraints}");
-                }
-                else
-                {
-                    sb.AppendLine($"    public static {builderName}<{allTypeParams}, TResult> {methodName}<{allTypeParams}, TResult, TKey>(");
-                    sb.AppendLine($"        this {thisBuilderName}<{allTypeParams}, TResult> builder,");
-                    sb.AppendLine($"        Expression<Func<{allTypeParams}, TKey>> _,");
-                    sb.AppendLine($"        Direction direction = Direction.Ascending) {constraints}");
-                }
-            }
-        }
-        else
-        {
-            if (keyType != null)
-            {
-                sb.AppendLine($"    public static {builderName}<{typeArgs}> {methodName}(");
-                sb.AppendLine($"        this {thisBuilderName}<{typeArgs}> builder,");
-                sb.AppendLine($"        Expression<Func<{typeArgs}, {keyType}>> _,");
-                sb.AppendLine($"        Direction direction = Direction.Ascending)");
-            }
-            else
-            {
-                var allTypeParams = string.Join(", ", Enumerable.Range(1, entityTypes.Count).Select(i => $"T{i}"));
-                var constraints = string.Join(" ", Enumerable.Range(1, entityTypes.Count).Select(i => $"where T{i} : class"));
-                sb.AppendLine($"    public static {builderName}<{allTypeParams}> {methodName}<{allTypeParams}, TKey>(");
-                sb.AppendLine($"        this {thisBuilderName}<{allTypeParams}> builder,");
-                sb.AppendLine($"        Expression<Func<{allTypeParams}, TKey>> _,");
-                sb.AppendLine($"        Direction direction = Direction.Ascending) {constraints}");
-            }
-        }
-
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.ResultTypeName != null && keyType != null)
-        {
-            var resultType = SanitizeTupleResultType(GetShortTypeName(site.ResultTypeName));
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{typeArgs}, {resultType}>>(builder);");
-        }
-        else if (site.ResultTypeName != null)
-        {
-            var allTypeParams = string.Join(", ", Enumerable.Range(1, entityTypes.Count).Select(i => $"T{i}"));
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{allTypeParams}, TResult>>(builder);");
-        }
-        else if (keyType != null)
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{typeArgs}>>(builder);");
-        }
-        else
-        {
-            var allTypeParams = string.Join(", ", Enumerable.Range(1, entityTypes.Count).Select(i => $"T{i}"));
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{allTypeParams}>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Simplified prebuilt chain path: passthrough (no SQL generation)
-        if (prebuiltChain != null)
-        {
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        {builderVar}.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            if (clauseBit.HasValue)
-            {
-                sb.AppendLine($"        {builderVar}.SetClauseBit({clauseBit.Value});");
-                sb.AppendLine($"        return builder;");
-            }
-            else
-            {
-                sb.AppendLine($"        return builder;");
-            }
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        if (clauseInfo is OrderByClauseInfo orderByInfo && orderByInfo.IsSuccess)
-        {
-            var escapedSql = EscapeStringLiteral(orderByInfo.ColumnSql);
-            var builderMethod = isOrderBy ? "AddOrderByClause" : "AddThenByClause";
-            sb.AppendLine($"        return {builderVar}.{builderMethod}(@\"{escapedSql}\", direction){bitSuffix};");
-        }
-        else if (clauseInfo != null && clauseInfo.IsSuccess)
-        {
-            var escapedSql = EscapeStringLiteral(clauseInfo.SqlFragment);
-            var builderMethod = isOrderBy ? "AddOrderByClause" : "AddThenByClause";
-            sb.AppendLine($"        return {builderVar}.{builderMethod}(@\"{escapedSql}\", direction){bitSuffix};");
-        }
-        else
-        {
-            // Non-translatable clause — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a Select() interceptor for joined query builders.
-    /// </summary>
-    private static void GenerateJoinedSelectInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName,
-        PrebuiltChainInfo? prebuiltChain = null, bool isFirstInChain = false)
-    {
-        var entityTypes = site.JoinedEntityTypeNames!.Select(GetShortTypeName).ToList();
-        var builderName = GetJoinedBuilderTypeName(entityTypes.Count);
-        var concreteBuilderName = ToConcreteTypeName(builderName);
-        var thisType = site.BuilderTypeName;
-        var thisBuilderName = builderName;
-        var typeArgs = string.Join(", ", entityTypes);
-        var projection = site.ProjectionInfo;
-
-        // Simplified prebuilt chain path: AsProjected instead of AddSelectClause
-        if (prebuiltChain != null && projection != null && projection.IsOptimalPath && projection.Columns.Count > 0)
-        {
-            var resultType = GetShortTypeName(projection.ResultTypeName);
-            sb.AppendLine($"    public static {builderName}<{typeArgs}, {resultType}> {methodName}(");
-            sb.AppendLine($"        this {thisBuilderName}<{typeArgs}> builder,");
-            sb.AppendLine($"        Func<{typeArgs}, {resultType}> _)");
-            sb.AppendLine($"    {{");
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{typeArgs}>>(builder);");
-            if (isFirstInChain && prebuiltChain.MaxParameterCount > 0)
-                sb.AppendLine($"        __b.AllocatePrebuiltParams({prebuiltChain.MaxParameterCount});");
-            sb.AppendLine($"        return __b.AsProjected<{resultType}>();");
-            sb.AppendLine($"    }}");
-            return;
-        }
-
-        if (projection != null && projection.IsOptimalPath && projection.Columns.Count > 0)
-        {
-            var resultType = GetShortTypeName(projection.ResultTypeName);
-            var columnNames = ReaderCodeGenerator.GenerateColumnNamesArray(projection);
-            var readerDelegate = ReaderCodeGenerator.GenerateReaderDelegate(projection, entityTypes[0]);
-
-            sb.AppendLine($"    public static {builderName}<{typeArgs}, {resultType}> {methodName}(");
-            sb.AppendLine($"        this {thisBuilderName}<{typeArgs}> builder,");
-            sb.AppendLine($"        Func<{typeArgs}, {resultType}> _)");
-            sb.AppendLine($"    {{");
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{typeArgs}>>(builder);");
-            sb.AppendLine($"        return __b.AddSelectClause<{resultType}>(");
-            sb.AppendLine($"            {columnNames},");
-            sb.AppendLine($"            {readerDelegate});");
-            sb.AppendLine($"    }}");
-        }
-        else
-        {
-            // Fallback
-            sb.AppendLine($"    public static {builderName}<{typeArgs}, TResult> {methodName}<TResult>(");
-            sb.AppendLine($"        this {thisBuilderName}<{typeArgs}> builder,");
-            sb.AppendLine($"        Func<{typeArgs}, TResult> selector)");
-            sb.AppendLine($"    {{");
-            sb.AppendLine($"        // Fallback - projection not fully analyzed at compile time");
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{typeArgs}>>(builder);");
-            sb.AppendLine($"        return __b.Select(selector);");
-            sb.AppendLine($"    }}");
-        }
-    }
-
-
-    /// <summary>
-    /// Generates an InsertBuilder ExecuteNonQueryAsync() interceptor.
-    /// </summary>
-    private static void GenerateInsertExecuteNonQueryInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var insertInfo = site.InsertInfo;
-
-        sb.AppendLine($"    public static Task<int> {methodName}(");
-        sb.AppendLine($"        this IInsertBuilder<{entityType}> builder,");
-        sb.AppendLine($"        CancellationToken cancellationToken = default)");
-        sb.AppendLine($"    {{");
-
-        if (insertInfo != null && insertInfo.Columns.Count > 0)
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<InsertBuilder<{entityType}>>(builder);");
-            // Generate column array
-            var columnNames = string.Join(", ", insertInfo.Columns.Select(c => $"@\"{EscapeStringLiteral(c.QuotedColumnName)}\""));
-            sb.AppendLine($"        // Set up columns (excluding identity and computed columns)");
-            sb.AppendLine($"        __b.SetColumns(new[] {{ {columnNames} }});");
-            sb.AppendLine();
-
-            // Generate entity property extraction
-            sb.AppendLine($"        // Extract values from each entity");
-            sb.AppendLine($"        foreach (var entity in __b.Entities)");
-            sb.AppendLine($"        {{");
-            sb.AppendLine($"            var paramIndices = new List<int>({insertInfo.Columns.Count});");
-
-            foreach (var column in insertInfo.Columns)
-            {
-                var valueExpr = column.IsForeignKey
-                    ? $"entity.{column.PropertyName}.Id"
-                    : $"entity.{column.PropertyName}";
-                if (column.CustomTypeMappingClass != null)
-                    valueExpr = $"{GetMappingFieldName(column.CustomTypeMappingClass)}.ToDb({valueExpr})";
-
-                var sensitiveArg = column.IsSensitive ? ", isSensitive: true" : "";
-                sb.AppendLine($"            paramIndices.Add(__b.AddParameter({valueExpr}{sensitiveArg}));");
-            }
-
-            sb.AppendLine($"            __b.AddRow(paramIndices);");
-            sb.AppendLine($"        }}");
-            sb.AppendLine();
-
-            sb.AppendLine($"        // Execute the insert");
-            sb.AppendLine($"        return Quarry.Internal.ModificationExecutor.ExecuteInsertNonQueryAsync(__b.State, __b.Entities, cancellationToken);");
-        }
-        else
-        {
-            // Non-translatable — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates an InsertBuilder ExecuteScalarAsync() interceptor for identity return.
-    /// </summary>
-    private static void GenerateInsertExecuteScalarInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var insertInfo = site.InsertInfo;
-
-        // ExecuteScalarAsync<TKey> is a generic method on generic class IInsertBuilder<T>.
-        // Interceptors must match the combined arity: <T, TKey> (CS9177).
-        // T is constrained to class to match IInsertBuilder<T> where T : class.
-        sb.AppendLine($"    public static Task<TKey> {methodName}<T, TKey>(");
-        sb.AppendLine($"        this IInsertBuilder<T> builder,");
-        sb.AppendLine($"        CancellationToken cancellationToken = default) where T : class");
-        sb.AppendLine($"    {{");
-
-        if (insertInfo != null && insertInfo.Columns.Count > 0)
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<InsertBuilder<T>>(builder);");
-            // Generate column array
-            var columnNames = string.Join(", ", insertInfo.Columns.Select(c => $"@\"{EscapeStringLiteral(c.QuotedColumnName)}\""));
-            sb.AppendLine($"        // Set up columns (excluding identity and computed columns)");
-            sb.AppendLine($"        __b.SetColumns(new[] {{ {columnNames} }});");
-            sb.AppendLine();
-
-            // Set identity column if present
-            if (!string.IsNullOrEmpty(insertInfo.IdentityColumnName))
-            {
-                sb.AppendLine($"        // Set identity column for RETURNING clause");
-                sb.AppendLine($"        __b.SetIdentityColumn(@\"{EscapeStringLiteral(insertInfo.IdentityColumnName!)}\");");
-                sb.AppendLine();
-            }
-
-            // Validate single entity
-            sb.AppendLine($"        // Validate single entity insert");
-            sb.AppendLine($"        if (__b.Entities.Count != 1)");
-            sb.AppendLine($"        {{");
-            sb.AppendLine($"            throw new InvalidOperationException(");
-            sb.AppendLine($"                \"ExecuteScalarAsync can only be used for single entity inserts. \" +");
-            sb.AppendLine($"                \"For batch inserts, use ExecuteNonQueryAsync() instead.\");");
-            sb.AppendLine($"        }}");
-            sb.AppendLine();
-
-            // Cast entity to concrete type for property access.
-            // At the call site T is always the concrete entity type (e.g. User),
-            // so this cast is safe.
-            sb.AppendLine($"        var entity = Unsafe.As<{entityType}>(__b.Entities[0]);");
-            sb.AppendLine($"        var paramIndices = new List<int>({insertInfo.Columns.Count});");
-
-            foreach (var column in insertInfo.Columns)
-            {
-                var valueExpr = column.IsForeignKey
-                    ? $"entity.{column.PropertyName}.Id"
-                    : $"entity.{column.PropertyName}";
-                if (column.CustomTypeMappingClass != null)
-                    valueExpr = $"{GetMappingFieldName(column.CustomTypeMappingClass)}.ToDb({valueExpr})";
-
-                var sensitiveArg = column.IsSensitive ? ", isSensitive: true" : "";
-                sb.AppendLine($"        paramIndices.Add(__b.AddParameter({valueExpr}{sensitiveArg}));");
-            }
-
-            sb.AppendLine($"        __b.AddRow(paramIndices);");
-            sb.AppendLine();
-
-            sb.AppendLine($"        // Execute the insert with identity return");
-            sb.AppendLine($"        return Quarry.Internal.ModificationExecutor.ExecuteInsertScalarAsync<T, TKey>(__b.State, __b.Entities[0], cancellationToken);");
-        }
-        else
-        {
-            // Non-translatable — skip interceptor entirely so the original method runs
-            return;
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates an InsertBuilder ToSql() interceptor that populates column metadata for SQL preview.
-    /// </summary>
-    private static void GenerateInsertToSqlInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-        var insertInfo = site.InsertInfo;
-
-        sb.AppendLine($"    public static string {methodName}(");
-        sb.AppendLine($"        this IInsertBuilder<{entityType}> builder)");
-        sb.AppendLine($"    {{");
-        sb.AppendLine($"        var __b = Unsafe.As<InsertBuilder<{entityType}>>(builder);");
-
-        if (insertInfo != null && insertInfo.Columns.Count > 0)
-        {
-            // Generate column array
-            var columnNames = string.Join(", ", insertInfo.Columns.Select(c => $"@\"{EscapeStringLiteral(c.QuotedColumnName)}\""));
-            sb.AppendLine($"        // Set up columns (excluding identity and computed columns)");
-            sb.AppendLine($"        __b.SetColumns(new[] {{ {columnNames} }});");
-            sb.AppendLine();
-
-            // Set identity column if present (for RETURNING/OUTPUT clause)
-            if (!string.IsNullOrEmpty(insertInfo.IdentityColumnName))
-            {
-                sb.AppendLine($"        // Set identity column for RETURNING clause");
-                sb.AppendLine($"        __b.SetIdentityColumn(@\"{EscapeStringLiteral(insertInfo.IdentityColumnName!)}\");");
-                sb.AppendLine();
-            }
-        }
-
-        sb.AppendLine($"        return __b.ToSqlDirect();");
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a placeholder interceptor for unknown kinds.
-    /// </summary>
-    /// <summary>
-    /// Generates a RawSqlAsync&lt;T&gt; interceptor with a typed reader delegate.
-    /// </summary>
-    private static void GenerateRawSqlAsyncInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var rawSqlInfo = site.RawSqlTypeInfo;
-        if (rawSqlInfo == null)
-            return;
-
-        var resultType = rawSqlInfo.ResultTypeName;
-
-        // Generate method signature matching the called overload
-        if (rawSqlInfo.HasCancellationToken)
-        {
-            sb.AppendLine($"    public static Task<List<{resultType}>> {methodName}(");
-            sb.AppendLine($"        this QuarryContext self,");
-            sb.AppendLine($"        string sql,");
-            sb.AppendLine($"        CancellationToken cancellationToken,");
-            sb.AppendLine($"        params object?[] parameters)");
-        }
-        else
-        {
-            sb.AppendLine($"    public static Task<List<{resultType}>> {methodName}(");
-            sb.AppendLine($"        this QuarryContext self,");
-            sb.AppendLine($"        string sql,");
-            sb.AppendLine($"        params object?[] parameters)");
-        }
-
-        sb.AppendLine($"    {{");
-
-        var ctArg = rawSqlInfo.HasCancellationToken ? "cancellationToken" : "CancellationToken.None";
-
-        if (rawSqlInfo.TypeKind == RawSqlTypeKind.Scalar)
-        {
-            // Scalar T: read reader.GetXxx(0) for each row
-            var readerMethod = rawSqlInfo.ScalarReaderMethod ?? "GetValue";
-            sb.AppendLine($"        return self.RawSqlAsyncWithReader(");
-            sb.AppendLine($"            sql,");
-            sb.AppendLine($"            static r => r.{readerMethod}(0),");
-            sb.AppendLine($"            {ctArg},");
-            sb.AppendLine($"            parameters);");
-        }
-        else
-        {
-            // Entity/DTO T: generate a switch-based reader
-            sb.AppendLine($"        return self.RawSqlAsyncWithReader(");
-            sb.AppendLine($"            sql,");
-            sb.AppendLine($"            static r =>");
-            sb.AppendLine($"            {{");
-            sb.AppendLine($"                var item = new {resultType}();");
-            sb.AppendLine($"                for (var i = 0; i < r.FieldCount; i++)");
-            sb.AppendLine($"                {{");
-            sb.AppendLine($"                    if (r.IsDBNull(i)) continue;");
-            sb.AppendLine($"                    switch (r.GetName(i))");
-            sb.AppendLine($"                    {{");
-
-            foreach (var prop in rawSqlInfo.Properties)
-            {
-                var assignment = GenerateRawSqlPropertyAssignment(prop);
-                sb.AppendLine($"                        case \"{prop.PropertyName}\": item.{prop.PropertyName} = {assignment}; break;");
-            }
-
-            sb.AppendLine($"                    }}");
-            sb.AppendLine($"                }}");
-            sb.AppendLine($"                return item;");
-            sb.AppendLine($"            }},");
-            sb.AppendLine($"            {ctArg},");
-            sb.AppendLine($"            parameters);");
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a RawSqlScalarAsync&lt;T&gt; interceptor with typed conversion.
-    /// </summary>
-    private static void GenerateRawSqlScalarAsyncInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var rawSqlInfo = site.RawSqlTypeInfo;
-        if (rawSqlInfo == null)
-            return;
-
-        var resultType = rawSqlInfo.ResultTypeName;
-
-        // Generate method signature matching the called overload
-        if (rawSqlInfo.HasCancellationToken)
-        {
-            sb.AppendLine($"    public static Task<{resultType}> {methodName}(");
-            sb.AppendLine($"        this QuarryContext self,");
-            sb.AppendLine($"        string sql,");
-            sb.AppendLine($"        CancellationToken cancellationToken,");
-            sb.AppendLine($"        params object?[] parameters)");
-        }
-        else
-        {
-            sb.AppendLine($"    public static Task<{resultType}> {methodName}(");
-            sb.AppendLine($"        this QuarryContext self,");
-            sb.AppendLine($"        string sql,");
-            sb.AppendLine($"        params object?[] parameters)");
-        }
-
-        sb.AppendLine($"    {{");
-
-        var ctArg = rawSqlInfo.HasCancellationToken ? "cancellationToken" : "CancellationToken.None";
-
-        // Generate typed converter
-        var converterExpr = GenerateScalarConverter(resultType);
-        sb.AppendLine($"        return self.RawSqlScalarAsyncWithConverter(");
-        sb.AppendLine($"            sql,");
-        sb.AppendLine($"            static v => {converterExpr},");
-        sb.AppendLine($"            {ctArg},");
-        sb.AppendLine($"            parameters);");
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a property assignment expression for a RawSql reader column case.
-    /// Handles enums, Ref&lt;&gt; FKs, custom type mappings, and standard types.
-    /// </summary>
-    private static string GenerateRawSqlPropertyAssignment(RawSqlPropertyInfo prop)
-    {
-        if (prop.CustomTypeMappingClass != null)
-        {
-            var dbReaderMethod = prop.DbReaderMethodName ?? "GetValue";
-            return $"new {prop.CustomTypeMappingClass}().FromDb(r.{dbReaderMethod}(i))";
-        }
-
-        if (prop.IsForeignKey && prop.ReferencedEntityName != null)
-        {
-            return $"new EntityRef<{prop.ReferencedEntityName}, {prop.ClrType}>(r.{prop.ReaderMethodName}(i))";
-        }
-
-        if (prop.IsEnum)
-        {
-            return $"({prop.FullClrType})r.{prop.ReaderMethodName}(i)";
-        }
-
-        return $"r.{prop.ReaderMethodName}(i)";
-    }
-
-    /// <summary>
-    /// Generates a converter expression for a scalar type from object to T.
-    /// Handles nullable, enum, and standard types.
-    /// </summary>
-    private static string GenerateScalarConverter(string resultType)
-    {
-        // Strip nullable for base type matching
-        var baseType = resultType.TrimEnd('?');
-        var isNullable = resultType.EndsWith("?");
-
-        return baseType switch
-        {
-            "int" or "Int32" => isNullable ? "(int?)Convert.ToInt32(v)" : "(int)Convert.ToInt64(v)",
-            "long" or "Int64" => isNullable ? "(long?)Convert.ToInt64(v)" : "Convert.ToInt64(v)",
-            "short" or "Int16" => isNullable ? "(short?)Convert.ToInt16(v)" : "(short)Convert.ToInt64(v)",
-            "byte" or "Byte" => isNullable ? "(byte?)(byte)Convert.ToInt64(v)" : "(byte)Convert.ToInt64(v)",
-            "bool" or "Boolean" => isNullable ? "(bool?)Convert.ToBoolean(v)" : "Convert.ToBoolean(v)",
-            "float" or "Single" => isNullable ? "(float?)Convert.ToSingle(v)" : "Convert.ToSingle(v)",
-            "double" or "Double" => isNullable ? "(double?)Convert.ToDouble(v)" : "Convert.ToDouble(v)",
-            "decimal" or "Decimal" => isNullable ? "(decimal?)Convert.ToDecimal(v)" : "Convert.ToDecimal(v)",
-            "string" or "String" => "v?.ToString() ?? string.Empty",
-            "Guid" => isNullable ? "(Guid?)(Guid)v" : "(Guid)v",
-            "DateTime" => isNullable ? "(DateTime?)Convert.ToDateTime(v)" : "Convert.ToDateTime(v)",
-            _ => $"({resultType})v"
-        };
-    }
-
-    private static void GeneratePlaceholderInterceptor(StringBuilder sb, UsageSiteInfo site, string methodName)
-    {
-        var entityType = GetShortTypeName(site.EntityTypeName);
-
-        sb.AppendLine($"    // TODO: Implement interceptor for {site.MethodName}");
-        sb.AppendLine($"    // file static ... {methodName}(...) {{ }}");
-    }
-
-    /// <summary>
-    /// Gets the DbDataReader method call for a specific column type and ordinal.
-    /// </summary>
-    private static string GetReaderCall(string clrType, int ordinal, bool isNullable)
-    {
-        // Handle nullable types - need null check
-        if (isNullable)
-        {
-            var baseType = GetNonNullableType(clrType);
-            var baseReaderMethod = GetReaderMethod(baseType);
-            return $"r.IsDBNull({ordinal}) ? default({clrType}) : r.{baseReaderMethod}({ordinal})";
-        }
-
-        var readerMethod = GetReaderMethod(clrType);
-        return $"r.{readerMethod}({ordinal})";
-    }
-
-    /// <summary>
-    /// Gets the appropriate DbDataReader method for a CLR type.
-    /// </summary>
-    private static string GetReaderMethod(string clrType)
-    {
-        return clrType switch
-        {
-            "bool" or "Boolean" or "System.Boolean" => "GetBoolean",
-            "byte" or "Byte" or "System.Byte" => "GetByte",
-            "sbyte" or "SByte" or "System.SByte" => "GetByte",
-            "short" or "Int16" or "System.Int16" => "GetInt16",
-            "ushort" or "UInt16" or "System.UInt16" => "GetInt16",
-            "int" or "Int32" or "System.Int32" => "GetInt32",
-            "uint" or "UInt32" or "System.UInt32" => "GetInt32",
-            "long" or "Int64" or "System.Int64" => "GetInt64",
-            "ulong" or "UInt64" or "System.UInt64" => "GetInt64",
-            "float" or "Single" or "System.Single" => "GetFloat",
-            "double" or "Double" or "System.Double" => "GetDouble",
-            "decimal" or "Decimal" or "System.Decimal" => "GetDecimal",
-            "string" or "String" or "System.String" => "GetString",
-            "char" or "Char" or "System.Char" => "GetChar",
-            "Guid" or "System.Guid" => "GetGuid",
-            "DateTime" or "System.DateTime" => "GetDateTime",
-            "DateTimeOffset" or "System.DateTimeOffset" => "GetValue",
-            "TimeSpan" or "System.TimeSpan" => "GetValue",
-            "DateOnly" or "System.DateOnly" => "GetValue",
-            "TimeOnly" or "System.TimeOnly" => "GetValue",
-            _ => "GetValue"
-        };
-    }
-
-    /// <summary>
-    /// Gets the non-nullable version of a nullable type.
-    /// </summary>
-    private static string GetNonNullableType(string clrType)
-    {
-        if (clrType.EndsWith("?"))
-        {
-            return clrType.Substring(0, clrType.Length - 1);
-        }
-
-        if (clrType.StartsWith("Nullable<") && clrType.EndsWith(">"))
-        {
-            return clrType.Substring(9, clrType.Length - 10);
-        }
-
-        if (clrType.StartsWith("System.Nullable<") && clrType.EndsWith(">"))
-        {
-            return clrType.Substring(16, clrType.Length - 17);
-        }
-
-        return clrType;
-    }
-
-    /// <summary>
-    /// Escapes a string for use in a verbatim string literal.
-    /// </summary>
-    private static string EscapeStringLiteral(string value)
-    {
-        return value.Replace("\"", "\"\"");
-    }
-
-    /// <summary>
-    /// Returns a <c>.WithClauseBit(N)</c> suffix for conditional clause interceptors,
-    /// or an empty string for unconditional clauses.
-    /// </summary>
-    private static string ClauseBitSuffix(int? bitIndex)
-        => bitIndex.HasValue ? $".WithClauseBit({bitIndex.Value})" : "";
-
-    // ───────────────────────────────────────────────────────────────
-    // Pre-built SQL Execution Interceptors (Tier 1)
-    // ───────────────────────────────────────────────────────────────
-
-    /// <summary>
-    /// Generates a tier 1 execution interceptor for SELECT queries (ExecuteFetchAll, ExecuteFetchFirst, etc.).
-    /// Contains a dispatch table that maps ClauseMask → pre-built SQL string literal.
-    /// </summary>
-    private static void GeneratePrebuiltSelectExecutionInterceptor(
-        StringBuilder sb,
-        UsageSiteInfo site,
-        string methodName,
-        PrebuiltChainInfo chain)
-    {
-        var entityType = GetShortTypeName(chain.EntityTypeName);
-
-        // Resolve result type using the same logic as the gate in GenerateInterceptorMethod.
-        // ResolveExecutionResultType returns the enriched type for tuples and single-column
-        // projections, which already has correct element types and names.
-        var rawResultType = ResolveExecutionResultType(
-            site.ResultTypeName, chain.ResultTypeName, chain.ProjectionInfo);
-
-        // Don't strip tuple element names — CS9148 requires exact match including names.
-        var resultType = !string.IsNullOrEmpty(rawResultType)
-            ? GetShortTypeName(rawResultType!)
-            : entityType;
-
-        // Determine return type and executor method from the execution kind.
-        // When the chain has parameters (MaxParameterCount > 0), use the prebuilt params variants
-        // which hydrate parameters from the pre-allocated array at terminal time.
-        var usePrebuiltParams = chain.MaxParameterCount > 0;
-        string returnType;
-        string executorMethod;
-        bool hasReader = true;
-
-        switch (site.Kind)
-        {
-            case InterceptorKind.ExecuteFetchAll:
-                returnType = $"Task<List<{resultType}>>";
-                executorMethod = usePrebuiltParams ? "ExecuteWithPrebuiltParamsAsync" : "ExecuteWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteFetchFirst:
-                returnType = $"Task<{resultType}>";
-                executorMethod = usePrebuiltParams ? "ExecuteFirstWithPrebuiltParamsAsync" : "ExecuteFirstWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteFetchFirstOrDefault:
-                returnType = $"Task<{resultType}?>";
-                executorMethod = usePrebuiltParams ? "ExecuteFirstOrDefaultWithPrebuiltParamsAsync" : "ExecuteFirstOrDefaultWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteFetchSingle:
-                returnType = $"Task<{resultType}>";
-                executorMethod = usePrebuiltParams ? "ExecuteSingleWithPrebuiltParamsAsync" : "ExecuteSingleWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteScalar:
-                // ExecuteScalar has a type parameter TScalar
-                returnType = $"Task<TScalar>";
-                executorMethod = usePrebuiltParams ? "ExecuteScalarWithPrebuiltParamsAsync<TScalar>" : "ExecuteScalarWithPrebuiltSqlAsync<TScalar>";
-                hasReader = false;
-                break;
-            case InterceptorKind.ToAsyncEnumerable:
-                returnType = $"IAsyncEnumerable<{resultType}>";
-                executorMethod = usePrebuiltParams ? "ToAsyncEnumerableWithPrebuiltParams" : "ToAsyncEnumerableWithPrebuiltSql";
-                break;
-            default:
-                return;
-        }
-
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-
-        // Method signature
-        if (site.Kind == InterceptorKind.ExecuteScalar)
-        {
-            // ExecuteScalar has arity 3: class-level TEntity+TResult + method-level TScalar.
-            // The interceptor must match this total arity.
-            sb.AppendLine($"    public static {returnType} {methodName}<TEntity, TResult, TScalar>(");
-            sb.AppendLine($"        this {thisType}<TEntity, TResult> builder,");
-            sb.AppendLine($"        CancellationToken cancellationToken = default) where TEntity : class");
-        }
-        else
-        {
-            sb.AppendLine($"    public static {returnType} {methodName}(");
-            sb.AppendLine($"        this {thisType}<{entityType}, {resultType}> builder,");
-            sb.AppendLine($"        CancellationToken cancellationToken = default)");
-        }
-
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.Kind == InterceptorKind.ExecuteScalar)
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<TEntity, TResult>>(builder);");
-        }
-        else
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteType}<{entityType}, {resultType}>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Dispatch table: switch on ClauseMask
-        GenerateDispatchTable(sb, chain.SqlMap, builderVar);
-
-        // Call the executor
-        if (hasReader && chain.ReaderDelegateCode != null)
-        {
-            sb.AppendLine($"        return {builderVar}.{executorMethod}(sql, {chain.ReaderDelegateCode}, cancellationToken);");
-        }
-        else if (hasReader)
-        {
-            // Fallback: no reader available — shouldn't happen for tier 1 SELECT, but be safe
-            sb.AppendLine($"        return {builderVar}.{executorMethod}(sql, cancellationToken);");
-        }
-        else
-        {
-            // ExecuteScalar: no reader delegate
-            sb.AppendLine($"        return {builderVar}.{executorMethod}(sql, cancellationToken);");
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a tier 1 execution interceptor for joined query execution (multi-entity SELECT).
-    /// </summary>
-    private static void GeneratePrebuiltJoinExecutionInterceptor(
-        StringBuilder sb,
-        UsageSiteInfo site,
-        string methodName,
-        PrebuiltChainInfo chain)
-    {
-        var joinedNames = chain.JoinedEntityTypeNames!;
-        var entityCount = joinedNames.Count;
-        var builderName = GetJoinedBuilderTypeName(entityCount);
-        var concreteBuilderName = ToConcreteTypeName(builderName);
-        var thisType = site.BuilderTypeName;
-        var thisBuilderName = builderName;
-        var entityTypes = joinedNames.Select(n => GetShortTypeName(n)).ToList();
-        var entityTypeArgs = string.Join(", ", entityTypes);
-
-        var rawResultType = ResolveExecutionResultType(
-            site.ResultTypeName, chain.ResultTypeName, chain.ProjectionInfo);
-        var resultType = !string.IsNullOrEmpty(rawResultType)
-            ? GetShortTypeName(rawResultType!)
-            : entityTypes[0];
-
-        var usePrebuiltParams = chain.MaxParameterCount > 0;
-        string returnType;
-        string executorMethod;
-        bool hasReader = true;
-
-        switch (site.Kind)
-        {
-            case InterceptorKind.ExecuteFetchAll:
-                returnType = $"Task<List<{resultType}>>";
-                executorMethod = usePrebuiltParams ? "ExecuteWithPrebuiltParamsAsync" : "ExecuteWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteFetchFirst:
-                returnType = $"Task<{resultType}>";
-                executorMethod = usePrebuiltParams ? "ExecuteFirstWithPrebuiltParamsAsync" : "ExecuteFirstWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteFetchFirstOrDefault:
-                returnType = $"Task<{resultType}?>";
-                executorMethod = usePrebuiltParams ? "ExecuteFirstOrDefaultWithPrebuiltParamsAsync" : "ExecuteFirstOrDefaultWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteFetchSingle:
-                returnType = $"Task<{resultType}>";
-                executorMethod = usePrebuiltParams ? "ExecuteSingleWithPrebuiltParamsAsync" : "ExecuteSingleWithPrebuiltSqlAsync";
-                break;
-            case InterceptorKind.ExecuteScalar:
-                returnType = $"Task<TScalar>";
-                executorMethod = usePrebuiltParams ? "ExecuteScalarWithPrebuiltParamsAsync<TScalar>" : "ExecuteScalarWithPrebuiltSqlAsync<TScalar>";
-                hasReader = false;
-                break;
-            case InterceptorKind.ToAsyncEnumerable:
-                returnType = $"IAsyncEnumerable<{resultType}>";
-                executorMethod = usePrebuiltParams ? "ToAsyncEnumerableWithPrebuiltParams" : "ToAsyncEnumerableWithPrebuiltSql";
-                break;
-            default:
-                return;
-        }
-
-        // Method signature with joined builder receiver type
-        if (site.Kind == InterceptorKind.ExecuteScalar)
-        {
-            sb.AppendLine($"    public static {returnType} {methodName}<{entityTypeArgs}, TResult, TScalar>(");
-            sb.AppendLine($"        this {thisBuilderName}<{entityTypeArgs}, TResult> builder,");
-            sb.AppendLine($"        CancellationToken cancellationToken = default) where TScalar : struct");
-        }
-        else
-        {
-            sb.AppendLine($"    public static {returnType} {methodName}(");
-            sb.AppendLine($"        this {thisBuilderName}<{entityTypeArgs}, {resultType}> builder,");
-            sb.AppendLine($"        CancellationToken cancellationToken = default)");
-        }
-
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        if (site.Kind == InterceptorKind.ExecuteScalar)
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{entityTypeArgs}, TResult>>(builder);");
-        }
-        else
-        {
-            sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderName}<{entityTypeArgs}, {resultType}>>(builder);");
-        }
-        var builderVar = "__b";
-
-        // Dispatch table: switch on ClauseMask
-        GenerateDispatchTable(sb, chain.SqlMap, builderVar);
-
-        // Call the executor
-        if (hasReader && chain.ReaderDelegateCode != null)
-        {
-            sb.AppendLine($"        return {builderVar}.{executorMethod}(sql, {chain.ReaderDelegateCode}, cancellationToken);");
-        }
-        else if (hasReader)
-        {
-            sb.AppendLine($"        return {builderVar}.{executorMethod}(sql, cancellationToken);");
-        }
-        else
-        {
-            sb.AppendLine($"        return {builderVar}.{executorMethod}(sql, cancellationToken);");
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a tier 1 execution interceptor for non-query operations (DELETE/UPDATE ExecuteNonQueryAsync).
-    /// </summary>
-    private static void GeneratePrebuiltNonQueryExecutionInterceptor(
-        StringBuilder sb,
-        UsageSiteInfo site,
-        string methodName,
-        PrebuiltChainInfo chain)
-    {
-        var entityType = GetShortTypeName(chain.EntityTypeName);
-
-        var thisType = site.BuilderTypeName;
-
-        // Determine builder type name based on query kind
-        string concreteBuilderTypeName;
-        string thisBuilderTypeName;
-        switch (chain.QueryKind)
-        {
-            case QueryKind.Delete:
-                concreteBuilderTypeName = $"ExecutableDeleteBuilder<{entityType}>";
-                thisBuilderTypeName = $"IExecutableDeleteBuilder<{entityType}>";
-                break;
-            case QueryKind.Update:
-                concreteBuilderTypeName = $"ExecutableUpdateBuilder<{entityType}>";
-                thisBuilderTypeName = $"IExecutableUpdateBuilder<{entityType}>";
-                break;
-            default:
-                return;
-        }
-
-        sb.AppendLine($"    public static Task<int> {methodName}(");
-        sb.AppendLine($"        this {thisBuilderTypeName} builder,");
-        sb.AppendLine($"        CancellationToken cancellationToken = default)");
-        sb.AppendLine($"    {{");
-
-        // Cast to concrete type (receiver is always an interface)
-        sb.AppendLine($"        var __b = Unsafe.As<{concreteBuilderTypeName}>(builder);");
-        var builderVar = "__b";
-
-        // Dispatch table: switch on ClauseMask
-        GenerateDispatchTable(sb, chain.SqlMap, builderVar);
-
-        sb.AppendLine($"        return {builderVar}.ExecuteWithPrebuiltSqlAsync(sql, cancellationToken);");
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates a tier 1 ToSql interceptor that returns the pre-built SQL string directly.
-    /// Handles SELECT, DELETE, and UPDATE chains, including joined queries.
-    /// </summary>
-    private static void GeneratePrebuiltToSqlInterceptor(
-        StringBuilder sb,
-        UsageSiteInfo site,
-        string methodName,
-        PrebuiltChainInfo chain)
-    {
-        var entityType = GetShortTypeName(chain.EntityTypeName);
-        var thisType = site.BuilderTypeName;
-        var concreteType = ToConcreteTypeName(thisType);
-
-        // Determine the full this-parameter type based on query kind and builder shape
-        string thisParamType;
-        string concreteParamType;
-
-        if (chain.IsJoinChain)
-        {
-            var joinedNames = chain.JoinedEntityTypeNames!;
-            var joinTypeArgs = string.Join(", ", joinedNames.Select(GetShortTypeName));
-
-            if (chain.ResultTypeName != null)
-            {
-                // Projected joined builder: IJoinedQueryBuilder<T1, T2, TResult>
-                var resultType = GetShortTypeName(
-                    ResolveExecutionResultType(site.ResultTypeName, chain.ResultTypeName, chain.ProjectionInfo)
-                    ?? chain.ResultTypeName);
-                thisParamType = $"{thisType}<{joinTypeArgs}, {resultType}>";
-                concreteParamType = $"{concreteType}<{joinTypeArgs}, {resultType}>";
-            }
-            else
-            {
-                // Unprojected joined builder: IJoinedQueryBuilder<T1, T2>
-                thisParamType = $"{thisType}<{joinTypeArgs}>";
-                concreteParamType = $"{concreteType}<{joinTypeArgs}>";
-            }
-        }
-        else if (chain.QueryKind == QueryKind.Select)
-        {
-            // SELECT: ToSql can be on IQueryBuilder<T> (unprojected) or IQueryBuilder<T, TResult> (projected)
-            if (chain.ResultTypeName != null)
-            {
-                var resultType = GetShortTypeName(
-                    ResolveExecutionResultType(site.ResultTypeName, chain.ResultTypeName, chain.ProjectionInfo)
-                    ?? chain.ResultTypeName);
-                thisParamType = $"{thisType}<{entityType}, {resultType}>";
-                concreteParamType = $"{concreteType}<{entityType}, {resultType}>";
-            }
-            else
-            {
-                thisParamType = $"{thisType}<{entityType}>";
-                concreteParamType = $"{concreteType}<{entityType}>";
-            }
-        }
-        else if (chain.QueryKind == QueryKind.Delete)
-        {
-            thisParamType = $"IExecutableDeleteBuilder<{entityType}>";
-            concreteParamType = $"ExecutableDeleteBuilder<{entityType}>";
-        }
-        else if (chain.QueryKind == QueryKind.Update)
-        {
-            thisParamType = $"IExecutableUpdateBuilder<{entityType}>";
-            concreteParamType = $"ExecutableUpdateBuilder<{entityType}>";
-        }
-        else
-        {
-            return;
-        }
-
-        sb.AppendLine($"    public static string {methodName}(");
-        sb.AppendLine($"        this {thisParamType} builder)");
-        sb.AppendLine($"    {{");
-
-        if (chain.SqlMap.Count == 1)
-        {
-            // Single variant — return const directly, no cast needed
-            foreach (var kvp in chain.SqlMap)
-            {
-                var escapedSql = EscapeStringLiteral(kvp.Value.Sql);
-                sb.AppendLine($"        return @\"{escapedSql}\";");
-            }
-        }
-        else
-        {
-            // Multiple variants — need concrete type for ClauseMask access
-            sb.AppendLine($"        return Unsafe.As<{concreteParamType}>(builder).ClauseMask switch");
-            sb.AppendLine($"        {{");
-            foreach (var kvp in chain.SqlMap.OrderBy(k => k.Key))
-            {
-                var escapedSql = EscapeStringLiteral(kvp.Value.Sql);
-                sb.AppendLine($"            {kvp.Key}UL => @\"{escapedSql}\",");
-            }
-            sb.AppendLine($"            _ => throw new InvalidOperationException(\"Unexpected ClauseMask\")");
-            sb.AppendLine($"        }};");
-        }
-
-        sb.AppendLine($"    }}");
-    }
-
-    /// <summary>
-    /// Generates the dispatch table switch expression that maps ClauseMask values
-    /// to pre-built SQL string literals.
-    /// </summary>
-    private static void GenerateDispatchTable(
-        StringBuilder sb,
-        Dictionary<ulong, PrebuiltSqlResult> sqlMap,
-        string builderVar = "builder")
-    {
-        Debug.Assert(sqlMap.Count > 0, "Dispatch table must have at least one SQL variant.");
-
-        if (sqlMap.Count == 1)
-        {
-            // Single variant — no switch needed, use const
-            foreach (var kvp in sqlMap)
-            {
-                var escapedSql = EscapeStringLiteral(kvp.Value.Sql);
-                sb.AppendLine($"        const string sql = @\"{escapedSql}\";");
-            }
-        }
-        else
-        {
-            // Multiple variants — switch expression on ClauseMask
-            sb.AppendLine($"        var sql = {builderVar}.ClauseMask switch");
-            sb.AppendLine($"        {{");
-
-            foreach (var kvp in sqlMap.OrderBy(k => k.Key))
-            {
-                var escapedSql = EscapeStringLiteral(kvp.Value.Sql);
-                sb.AppendLine($"            {kvp.Key}UL => @\"{escapedSql}\",");
-            }
-
-            sb.AppendLine($"            _ => throw new InvalidOperationException(\"Unexpected ClauseMask: \" + {builderVar}.ClauseMask)");
-            sb.AppendLine($"        }};");
-        }
-    }
-
-    /// <summary>
-    /// Returns the concrete type name corresponding to a concrete or interface builder type name.
-    /// </summary>
-    private static string ToConcreteTypeName(string typeName)
-    {
-        if (typeName.Length > 1 && typeName[0] == 'I' && char.IsUpper(typeName[1]))
-            return typeName.Substring(1);
-        return typeName;
-    }
-
-    /// <summary>
-    /// Gets a short type name from a fully qualified type name.
-    /// </summary>
-    internal static string GetShortTypeName(string fullTypeName)
-    {
-        // Remove global:: prefix
-        if (fullTypeName.StartsWith("global::"))
-        {
-            fullTypeName = fullTypeName.Substring(8);
-        }
-
-        // For now, return as-is. In later phases, we may want to shorten common types.
-        return fullTypeName;
     }
 }
